### 2023-05-04 #110 @slobentanzer
BioCypher submission
Submitting Author: Sebastian Lobentanzer (@slobentanzer)
Package Name: biocypher
One-Line Description of Package: framework for creating biomedical knowledge graphs
Repository Link: https://github.com/biocypher/biocypher
Version submitted: 0.5.14
Reviewer 2: @Midnighter (with Sinna)
Version accepted: TBD

---

BioCypher is a flexible and modular framework for the creation of knowledge graphs on the basis of ontologies, yielding a number of representations of the input knowledge sources. Its aim is to increase accessibility of biomedical knowledge graphs to a wide range of biomedical researchers.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.
