### 2023-05-11 bdccf1b
update tutorial, include pandas output version .. (additionally, suffixed)
