+++ tutorial/03__implicit_subclass_pandas.py
@@ -1,10 +1,15 @@
 from biocypher import BioCypher
-from tutorial.data_generator import Protein
+from tutorial.data_generator import Protein, EntrezProtein
 
 
 def main():
     # Setup: create a list of proteins to be imported
-    proteins = [Protein() for _ in range(10)]
+    proteins = [
+        p for sublist in zip(
+            [Protein() for _ in range(10)],
+            [EntrezProtein() for _ in range(10)],
+        ) for p in sublist
+    ]
 
     # Extract id, label, and property dictionary
     def node_generator():
@@ -17,12 +22,16 @@ def main():
 
     # Create BioCypher driver
     bc = BioCypher(
-        biocypher_config_path='tutorial/01_biocypher_config.yaml',
-        schema_config_path='tutorial/01_schema_config.yaml',
+        biocypher_config_path='tutorial/03_biocypher_config.yaml',
+        schema_config_path='tutorial/03_schema_config.yaml',
     )
     # Run the import
-    nodes = bc.to_df(node_generator())
-    print(nodes)
+    bc.add(node_generator())
+
+    for name, df in bc.to_df().items():
+        print(name)
+        print(df)
+
 
 if __name__ == '__main__':
     main()
