+++ docs/tutorial.md
@@ -13,15 +13,20 @@ interface - are represented by simulated data containing some examples of
 differently formatted biomedical entities such as proteins and their
 interactions.
 
+There are two versions of this tutorial, which only differ in the output format.
+The first uses a CSV output format to write files suitable for Neo4j admin
+import, and the second creates an in-memory collection of Pandas dataframes.
+You can find both in the `tutorial` directory of the BioCypher repository; the
+Pandas version of each tutorial step is suffixed with `_pandas`.
+
 ```{admonition} Neo4j
 :class: warning
 
-We use the BioCypher output adapter to write CSV files meant for Neo4j admin
-import, mainly because it is the use case with the lowest amount of
-dependencies.  While you can use the files to create an actual Neo4j database,
-it is not required for this tutorial. For checking the output, you can simply
-open the CSV files in a text editor or your IDE; by default, they will be
-written to the ``biocypher-out`` directory.
+While you can use the files generated to create an actual Neo4j database, it is
+not required for this tutorial. For checking the output, you can simply open the
+CSV files in a text editor or your IDE; by default, they will be written to the
+``biocypher-out`` directory. If you simply want to run the tutorial to see how
+it works, you can also run the Pandas version.
 
 ```
 
