+++ docs/index.rst
@@ -81,7 +81,8 @@ We aim to **enable access to versatile and powerful knowledge graphs for as many
 researchers as possible**. Making biomedical knowledge "their own" is often a
 privilege of the companies and groups that can afford individuals or teams
 working on knowledge representation in a dedicated manner. With BioCypher, we
-aim to change that. To achieve this, we have developed a framework that
+aim to change that. Creating a knowledge graph should be "as simple as possible,
+but not any simpler." To achieve this, we have developed a framework that
 facilitates the creation of knowledge graphs that are informed by the latest
 developments in the field of biomedical knowledge representation. However, to
 make this framework truly accessible and comprehensive, we need the input of the
