### 2023-05-12 #110 @lwasser
hi @slobentanzer thank you for this submission and welcome to pyOpenSci! I just wanted to say hello 👋  and to let you know that we see this submission. We will get back to you with the start of our initial checks next week! In the meantime, have a wonderful weekend. we will be in touch. ✨
