### 2023-05-15 d517f4e
add bioconductor pointer in docs
