+++ docs/contents.rst
@@ -12,4 +12,5 @@
    tutorial-ontology
    tutorial-adapter
    neo4j
+   r-bioc
    submodule
