+++ _sources/r-bioc.md.txt
@@ -1,0 +1,5 @@
+# R and Bioconductor
+We are working on a Bioconductor package to make BioCypher functionality
+available to the R community. The current work in progess is available in [this
+repository](https://vjcitn.github.io/biocBiocypher/index.html). If you are
+interested in contributing or using the package, please get in touch!
