### 2023-05-15 608b6ed
remove id properties as reserved, debug log
