### 2023-05-15 f4ef0af
relationship ID becomes a permanent part of the .. batch writer output files (in second place column after the source ID) `id` for neo4j, `_ID` for postgres, `_key` for arango to stay consistent with nodes (did not check whether arango prescribes a keyword) adjusted tests to check this behaviour and still pass
