+++ test/test_write_neo4j.py
@@ -471,11 +471,31 @@ def test_write_edge_data_from_gen(bw, _get_edges):
         c = f.read()
 
     assert passed
-    assert "p0;'T253';4;p1;PERTURBED_IN_DISEASE" in l
-    assert "p1;'T253';4;p2;PERTURBED_IN_DISEASE" in l
+    assert "p0;" in l
+    assert "prel0;" in l
+    assert "'T253';" in l
+    assert "4;" in l
+    assert "p1;" in l
+    assert "PERTURBED_IN_DISEASE" in l
+    assert "p1;" in l
+    assert "prel1;" in l
+    assert "'T253';" in l
+    assert "4;" in l
+    assert "p2;" in l
+    assert "PERTURBED_IN_DISEASE" in l
     assert '\n' in l
-    assert "m0;'3-UTR';1;p1;Is_Mutated_In" in c
-    assert "m1;'3-UTR';1;p2;Is_Mutated_In" in c
+    assert "m0;" in c
+    assert "mrel0;" in c
+    assert "'3-UTR';" in c
+    assert "1;" in c
+    assert "p1;" in c
+    assert "Is_Mutated_In" in c
+    assert "m1;" in c
+    assert "mrel1;" in c
+    assert "'3-UTR';" in c
+    assert "1;" in c
+    assert "p2;" in c
+    assert "Is_Mutated_In" in c
     assert '\n' in c
 
 
@@ -524,12 +544,23 @@ def test_write_edge_data_from_list(bw, _get_edges):
         c = f.read()
 
     assert passed
-    assert "p0;'T253';4;p1;PERTURBED_IN_DISEASE" in l
-    assert "p1;'T253';4;p2;PERTURBED_IN_DISEASE" in l
-    assert '\n' in l
-    assert "m0;'3-UTR';1;p1;Is_Mutated_In" in c
-    assert "m1;'3-UTR';1;p2;Is_Mutated_In" in c
+    assert "p0;" in l
+    assert "prel0;" in l
+    assert "'T253';" in l
+    assert "4;" in l 
+    assert "p1;" in l
+    assert "PERTURBED_IN_DISEASE" in l
+    assert "\n" in l
+    assert "p2;PERTURBED_IN_DISEASE" in l
+    assert "m0;" in c
+    assert "mrel0;" in c
+    assert "'3-UTR';" in c
+    assert "1;" in c
+    assert "p1;" in c
+    assert "Is_Mutated_In" in c
+    assert "m1;" in c
     assert '\n' in c
+    
 
 
 def test_write_edge_data_from_list_no_props(bw):
@@ -562,11 +593,19 @@ def test_write_edge_data_from_list_no_props(bw):
         c = f.read()
 
     assert passed
-    assert 'p0;p1;PERTURBED_IN_DISEASE' in l
-    assert 'p1;p2;PERTURBED_IN_DISEASE' in l
+    assert 'p0;' in l
+    assert 'p1;' in l
+    assert 'PERTURBED_IN_DISEASE' in l
+    assert 'p1;' in l
+    assert 'p2;' in l
+    assert 'PERTURBED_IN_DISEASE' in l
     assert '\n' in l
-    assert 'm0;p1;Is_Mutated_In' in c
-    assert 'm1;p2;Is_Mutated_In' in c
+    assert 'm0;' in c
+    assert 'p1;' in c
+    assert 'Is_Mutated_In' in c
+    assert 'm1;' in c
+    assert 'p2;' in c
+    assert 'Is_Mutated_In' in c
     assert '\n' in c
 
 
@@ -606,8 +645,8 @@ def test_write_edge_data_headers_import_call(bw, _get_nodes, _get_edges):
     with open(call_csv) as f:
[…]
