+++ biocypher/_metadata.py
@@ -19,7 +19,7 @@ import importlib.metadata
 
 import toml
 
-_VERSION = '0.5.14'
+_VERSION = '0.5.15'
 
 
 def get_metadata():
