+++ .bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.5.14
+current_version = 0.5.15
 commit = True
 tag = True
 files = pyproject.toml
