### 2023-05-16 #110 @NickleDave
Welcome to pyOpenSci @slobentanzer and thank you for your patience.

I'm happy to report that biocypher passes almost all of the initial editor checks.

There is one thing we need you to take care of before we start the review:
>  - [ ] **README** The package has a `README.md` file with clear explanation of what the package does, instructions on how to install it, and a link to development instructions.

Please add the following to the README:
- [ ] a **clear** explanation of what the package does, such that it could be understood by a general scientific audience.
  - What is a knowledge graph? How are they used? A link to a review or wikipedia article can supplement,
  - It would really help to have a precise, succinct definition of what the package does and how it does it--use verbs--that is understandable even to people that don't already work with knowledge graphs. "Allows users to construct knowledge graphs from heterogeneous biomedical data, using adapters written in Python". This should ideally be the first thing in the README. The description that comes a bit lower has a helpful figure, but the language makes it sound very abstract without any detail on what or how, "this is our proposal for a framework..."
- [ ] instructions on how to install: this can be simple, `pip install biocypher` and if there is a package on a community conda channel, `conda install biocypher -c conda-forge` (or bioconda or whatever)

Please let me know once you have addressed this requirement.

The filled-out template with checks follows.

Please check our [Python packaging guide](https://www.pyopensci.org/python-package-guide) for more information on the elements below.

    - `pip install biocypher` works. There is not a conda-forge package as far as I can tell
  - [x] The package imports properly into a standard Python environment `import package-name`.
  - [x] API documentation (documentation for your code's functions, classes, methods and attributes): this includes clearly written docstrings with variables defined using a standard docstring format. *We suggest using the [Numpy](https://numpydoc.readthedocs.io/en/latest/format.html#docstring-standard) docstring format*.
    - There is a README but it does not have an explanation of what the package does or install instructions8
  - [x] **Code of Conduct** The package has a `Code of Conduct` file.
- [x] **Automated tests** Package has a testing suite and is tested via GitHub actions or another Continuous Integration service.
  - there is already a pre-print associated with the package, see link in README

---
---

None of these are required but are just things I noted during the checks:
- In addition to adding the required clear explanation and install instructions to the README, you might consider adding more info from the index of the docs; there is a lot of very useful info / context there. For more on good practices for READMEs, see https://www.pyopensci.org/python-package-guide/documentation/repository-files/readme-file-best-practices.html
- It's a bit confusing that the [install instructions](https://biocypher.org/installation.html#installation) (1) recommend a specific development tool and show adding the package as a dependency with that tool, and (2) use pip inside a conda environment. Typically one might show `pip install package` for the first and `conda install package -c conda-forge` for the second
- I noticed the package defaults to logging at the INFO level and writing a log file on import. Not sure that everyone will want that by default, especially if they are on a shared system
