+++ README.md
@@ -1,18 +1,15 @@
 # BioCypher
 
-## 📖 Documentation
-Tutorial and developer docs at https://biocypher.org. For a quickstart into your
-own pipeline, you can refer to our [project
-template](https://github.com/biocypher/project-template), and for an
-overview of existing and planned adapters for resources and outputs, as well
-as other features, visit our [GitHub Project
-Board](https://github.com/orgs/biocypher/projects/3/views/2).
-
 ## ❓ Description
-This is the development repository for BioCypher, our proposal for a unifying
-framework to create knowledge graph databases for systems biomedicine. For an
-overview, usage notes, and a tutorial, read the docs
-[here](https://biocypher.org).
+Knowledge graphs (KGs) are an [approach to knowledge
+representation](https://en.wikipedia.org/wiki/Knowledge_graph) that uses graph
+structure to facilitate exploration and analysis of complex data, often
+leveraging semantic information. They are popular in many research areas,
+including the life sciences, due to their versatile use, for instance in data
+storage, integration, reasoning, and more recently in artificial intelligence.
+The creation of KGs is a complex task; BioCypher helps you in creating and
+maintaining your own KG. For more overview, usage notes, and a tutorial, read
+the docs [here](https://biocypher.org).
 
 <img
     style="display: block;
@@ -23,20 +20,30 @@ overview, usage notes, and a tutorial, read the docs
     alt="Graphical Abstract">
 </img>
 
-## ⚙️ Usage
-BioCypher is currently in prototype stage. Installation instructions can be
-found [here](https://biocypher.org/installation.html).
-
-Exemplary usage of BioCypher to build a graph database is shown in various
-pipelines we have created. You can find these on the [Components Project
+## 📖 Documentation
+Tutorial and developer docs at https://biocypher.org. For a quickstart into your
+own pipeline, you can refer to our [project
+template](https://github.com/biocypher/project-template), and for an overview of
+existing and planned adapters for resources and outputs, as well as other
+features, visit our [GitHub Project
 Board](https://github.com/orgs/biocypher/projects/3/views/2).
+
+## ⚙️ Installation / Usage
+Install the package from PyPI using `pip install biocypher`. More comprehensive
[…]
