### 2023-05-17 #110 @slobentanzer
Hi @NickleDave,
thanks for the feedback, I added the requested details to the README.

Regarding your comments, here's my opinion:
- I prefer not having dual maintenance of content, i.e., I do not want to keep a readme and the docs up to date simultaneously; so I am not in favour of packing things from the docs into the readme, because they need to be kept in sync in this case. What is the benefit of having it in the README, if users can just read the docs? BTW, that was also the reason that the docs were the first section of the README, as opposed to the description as suggested by you.
- I always use pip inside conda envs and did not know this was unusual. Happy for suggestions as to how to make that section clearer or more aligned with good practise.
- I am new to Python and find logging quite confusing; happy for suggestions (or a pointer to a best practise resource) here as well.
