### 2023-05-17 f9bd548
fix error for node w/o simple edges
