### 2023-05-22 #110 @NickleDave
Thank you @slobentanzer for adding those details to the README. Looks great!

> What is the benefit of having it in the README, if users can just read the docs?

I hear you. I also do not like the idea of duplicating content in the README that can go out of sync.
One way around this is to have the single source of truth be your README, and then get that into your docs by using an [include directive in RST](https://docutils.sourceforge.io/docs/ref/rst/directives.html#include) or in [MyST](https://myst-parser.readthedocs.io/en/latest/faq/index.html#include-rst-files-into-a-markdown-file)

> What is the benefit of having it in the README, if users can just read the docs?

There is an argument to be made that developers hyper-fixate on "Awesome" READMEs, mainly to impress other developers. Still, you don't want to lose a user because they found your GitHub repo first, and they couldn't figure out what your project really does, so they gave up without bothering to read the docs. What is the killer feature and what does it look like in code? If you can express that in one sentence and a single snippet with a few lines of code, at the top of both your README and on the index page of your docs, then you know you're making the best case for your package to new users. Of course that's not possible for every package, but seems like a good general goal to aim for.

> I always use pip inside conda envs and did not know this was unusual. Happy for suggestions as to how to make that section clearer or more aligned with good practise.

I am not a conda expert but my understanding is that it's good practice to prefer `conda install` over `pip install`, to be able to reproduce environments. Basically because pip does things that conda doesn't do, or that conda doesn't know about. Maybe @jjhelmus can let us know if the advice in this blog post is still up-to-date:
https://www.anaconda.com/blog/using-pip-in-a-conda-environment

> I am new to Python and find logging quite confusing; happy for suggestions (or a pointer to a best practise resource) here as well.

I am less new to Python but I still find logging confusing 😭

I'd suggest a quote from Hitchhiker's guide to Python is relevant here:

> the user, not the library, should dictate what happens when a logging event occurs

https://docs.python-guide.org/writing/logging/#logging-in-a-library

For that reason I would avoid forcing a logging file and messages on your users.

I have looked several times but have never found a guide to logging that I felt answered my questions.

You probably know these already but:
* the HOWTO in the standard library docs is a good place to start, if you haven't read it already: https://docs.python.org/3/howto/logging.html
* Real Python usually has friendly introductions to topics, and logging is no exception: https://realpython.com/python-logging/
* If that's not enough, here is a rather opinionated blog post: https://ppwwyyxx.com/blog/2022/Effective-Python-Logging/

The main question I had as a library developer is how to use the logger in my modules, and how should I configure it. Through trial and error I have found it's "good enough" practice to make a new logger instance in each module, assigning it `__name__` as its name (as you will read in most guides). I have also found that if I *do* want to configure those logger instances, it's easiest to achieve this through a function I call that checks if there are already existing Handlers attached, so that I don't stomp on a user's existing configuration. See here: https://github.com/vocalpy/vak/pull/535

In my experience is that most scientific Python libraries do not include logging. There are exceptions I can think of:
* [pytorch-lightning](https://lightning.ai/docs/pytorch/latest/)
* [armory](https://github.com/twosixlabs/armory)
* [scikit-learn](https://github.com/scikit-learn/scikit-learn) and see discussion on https://github.com/scikit-learn/scikit-learn/issues/17439)

You might find reading the code of those libraries helpful
