### 2023-05-31 bbf0743
small tweaks to the jpynb and added ipykernel
