+++ pyproject.toml
@@ -52,6 +52,7 @@ pytest-cov = "^3.0.0"
 hypothesis = "^6.50.1"
 isort = "^5.10.1"
 ipython = "^8.7.0"
+ipykernel = "^6.23.1"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
