### 2023-06-01 #110 @arianesasso
Hello @slobentanzer 👋🏼 . Thanks for the intro @NickleDave :). It is my pleasure to review **BioCypher**! We are now in the process of finding reviewers for the package. Hopefully, depending on their response, this should take one to two weeks. I will keep you posted on this thread, and if you have any questions, feel free to reach out!
