+++ biocypher/_config/test_schema_config.yaml
@@ -102,6 +102,7 @@ post translational interaction:
 phosphorylation:
   is_a: post translational interaction
   represented_as: edge
+  use_id: false
   input_label: phosphorylation
 
 gene to disease association:
@@ -134,7 +135,7 @@ gene to gene association:
     score: float
     id: str  # should be removed
 
-gene to variant association:
+gene to variant association:  # should be removed
   is_a: gene to variant association
   represented_as: edge
   input_label: gene_variant
