### 2023-06-12 a33fa20
conditional adjustment in edge data writing
