### 2023-06-12 57b1726
fix behaviour for RelAsNode type
