+++ test/test_write_neo4j.py
@@ -575,21 +575,32 @@ def test_write_edge_id_optional(bw, _get_edges):
         )
         edges.append(e1)
 
-    passed = bw._write_edge_data(edges, batch_size=int(1e4))
+    passed = bw.write_edges(edges, batch_size=int(1e4))
     assert passed
 
     tmp_path = bw.outdir
 
-    apl_csv = os.path.join(tmp_path, 'PERTURBED_IN_DISEASE-part000.csv')
-    pps_csv = os.path.join(tmp_path, 'Phosphorylation-part000.csv')
+    pert_csv = os.path.join(tmp_path, 'PERTURBED_IN_DISEASE-part000.csv')
+    phos_csv = os.path.join(tmp_path, 'Phosphorylation-part000.csv')
 
-    with open(apl_csv) as f:
-        l = f.read()
-    with open(pps_csv) as f:
-        c = f.read()
+    with open(pert_csv) as f:
+        pertf = f.read()
+    with open(phos_csv) as f:
+        phosf = f.read()
 
-    assert "prel0;" in l
-    assert "phos1" not in c
+    assert "prel0;" in pertf
+    assert "phos1;" not in phosf
+
+    pert_header = os.path.join(tmp_path, 'PERTURBED_IN_DISEASE-header.csv')
+    phos_header = os.path.join(tmp_path, 'Phosphorylation-header.csv')
+
+    with open(pert_header) as f:
+        perth = f.read()
+    with open(phos_header) as f:
+        phosh = f.read()
+
+    assert "id;" in perth
+    assert "id;" not in phosh
 
 def test_write_edge_data_from_list_no_props(bw):
     le = 4
