+++ docs/conf.py
@@ -82,3 +82,7 @@ html_theme_options = {
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 # html_static_path = ['_static']
+
+# -- OpenGraph configuration -------------------------------------------------
+
+ogp_site_url = 'https://biocypher.org'
