### 2023-06-16 c679983
put tags in index.rst instead of conf
