+++ _sources/index.rst.txt
@@ -1,5 +1,5 @@
-:og:title: BioCypher
-:og:description: A framework for creating biomedical knowledge graphs.
+:og:title: BioCypher - Democratising Knowledge Graphs
+:og:description: BioCypher is a framework to simplify the creation of robust biomedical knowledge graphs.
 
 .. image:: banner.png
 
