+++ docs/conf.py
@@ -72,7 +72,7 @@ autodoc_mock_imports = ['bmt', 'neo4j-utils']
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 
-html_title = 'BioCypher - Democratising Knowledge Graphs'
+html_title = 'BioCypher'
 html_theme = 'sphinx_rtd_theme'
 html_theme_options = {
     'navigation_depth': 2,
@@ -91,7 +91,7 @@ ogp_image = 'https://biocypher.org/_images/biocypher-open-graph.png'
 ogp_custom_meta_tags = [
     '<meta property="twitter:card" content="summary_large_image" />',
     '<meta property="twitter:url" content="https://biocypher.org" />',
-    '<meta property="twitter:title" content="BioCypher - Democratising Knowledge Graphs" />',
+    '<meta property="twitter:title" content="Democratising Knowledge Graphs - BioCypher" />',
     '<meta property="twitter:description" content="BioCypher is a framework to simplify the creation of robust biomedical knowledge graphs." />',
     '<meta property="twitter:image" content="https://biocypher.org/_images/biocypher-open-graph.png" />',
 ]
