### 2023-06-20 2acc798
add implicit inheritance tests
