+++ .bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.5.16
+current_version = 0.5.17
 commit = True
 tag = True
 files = pyproject.toml
