+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "biocypher"
-version = "0.5.16"
+version = "0.5.17"
 description = "A unifying framework for biomedical research knowledge graphs"
 authors = [
     "Sebastian Lobentanzer <sebastian.lobentanzer@gmail.com>",
