### 2023-06-22 f426463
First minimal working version for rdf conversion
