### 2023-07-08 #110 @arianesasso
@slobentanzer Great! Sorry for the delay, but now we have reviewers for BioCypher 🥳 .

@Zethson will be our first reviewer and should be starting his review soon.

@Zethson Thank you so much for your contribution! Please, check our [reviewer's guide](https://www.pyopensci.org/peer-review-guide/software-peer-review-guide/reviewer-guide.html), the reviewer template, and the Python packaging standards detailed in the [packaging guide](https://www.pyopensci.org/python-package-guide). Also, please reach out in case you have any questions. Happy reviewing!
