### 2023-07-13 #110 @arianesasso
Dear @slobentanzer, we now have our second reviewers 🥳 !

@Midnighter will be the second reviewer, together with his colleague Sinna. She will cover the domain knowledge while he will look more on the Python side. We ask you for a bit of patience since now people are mainly on Summer vacation, so the review might take a bit longer.

@Midnighter Thank you so much to both of you for agreeing to review BioCypher! Please, check our [reviewer's guide](https://www.pyopensci.org/peer-review-guide/software-peer-review-guide/reviewer-guide.html), the reviewer template, and the Python packaging standards detailed in the [packaging guide](https://www.pyopensci.org/python-package-guide). Also, please don't hesitate to reach out if you have any questions. Happy reviewing!
