### 2023-07-13 #110 @arianesasso
@Zethson @Midnighter (and Sinna), before starting to review, please fill out our [pre-review survey](https://forms.gle/F9mou7S3jhe8DMJ16). This helps us improve all aspects of our review and better understand our community. No personal data will be shared from this survey - it will only be used in an aggregated format by our Executive Director to improve our processes and programs.

- [ ] @Zethson survey completed.
- [ ] @Midnighter survey completed.
- [ ] Sinna survey completed.

here as a comment once your review is complete.

Please get in touch with any questions or concerns! Your review is due: **15 August 2023**.

---

Reviewers: @Zethson and @Midnighter (and Sinna)
Due date: **15 August 2023**
