+++ _sources/index.rst.txt
@@ -17,7 +17,7 @@ biomedical community.
 
 If you're new to knowledge graphs and want to familiarise with the concepts that
 drive BioCypher, we recommend to check out the graphical abstract below and read
-[our paper](https://www.nature.com/articles/s41587-023-01848-y)!
+`our paper <https://www.nature.com/articles/s41587-023-01848-y>`_!
 
 .. grid:: 2
    :gutter: 2
