### 2023-07-21 #110 @Zethson
## Lukas (zethson) review

- [X] As the reviewer I confirm that there are no conflicts of interest for me to review this work.

---

Hey @slobentanzer !

Really amazing work with Biocypher and it really shows that you're trying hard and succeeding with creating a package that is maintainable, well documented and works. I have a few random comments that you can consider addressing:

1. https://biocypher.org/tutorial.html suggests to install biocypher from source with poetry. I think that this is "noise" here and that all tutorials should always work with the latest version on pypi.
2. Nit: On the tutorial site terms like "DBMS instance" are not introduced.
3. Am I wrong or are the Pandas versions of the tutorials not rendered?
4. API documentation is somewhat "hidden" under "submodule documentation". "API documentation" or something of that sort is probably more common. I also feel like it mostly shows private functions (based on underscores) but not really public ones?
5. I'm having trouble finding a CI that runs all tests. Am I looking at the wrong places?
