### 2023-07-21 biocypher/biocypher#218 @slobentanzer
CI/CD: create workflow for running tests automatically on push
One big question is: how do we run tests with DBMS dependencies? Connect to docker containers of a compose that is installed on the CI server?
