### 2023-07-21 biocypher/biocypher#219 @slobentanzer
Documentation: overhaul quickstart
Simplest way of showing:
- the high-level process from inception to build of KG
- explanation and examples of all user-facing functions
