### 2023-07-21 #110 @slobentanzer
Hi @Zethson, huge thanks for taking the time to review the package! It is accurate. I will try to address the missing points and comments as time allows. Here are some explanations, feel free to suggest ways to improve the current state:

1. The reason we recommend to clone and install the package locally is not because of code dependence, but rather because the tutorial files are included in the repo, but not the package. Do you (or another person reading this thread) have another solution for making the tutorial code available to the user in the simplest way?
2. Will go over the docs and [explain all abbreviations](https://github.com/biocypher/biocypher/issues/217) soon.
3. For me it is. (The first couple functions for setup do not have much of an output.) Could be something to do with how the colab works (each user has their own version, maybe)? I don't usually work with colab, so feel free to suggest improvements.
4. I am using the Sphinx autodoc module for rendering the function documentation. There is already an [issue](https://github.com/biocypher/biocypher/issues/183) for making it work for all modules, but I have not had time to delve deep into Sphinx functionality to figure this out. We do only have a fairly small amount of user-facing functions in the first place, which are explained in the tutorials. I could maybe add a manual explanation of the most important functions in [quickstart](https://github.com/biocypher/biocypher/issues/219), but that would not be automatic any more.
5. You are right, there is no action for CI. I have been doing it manually so far due to lack of time to establish a workflow, but I agree that this takes high priority. One issue is that some tests depend on DBMSs running on the test machine, so this is not a straightforward setup. It is on [the list](https://github.com/biocypher/biocypher/issues/218).
