### 2023-07-21 #110 @Zethson
> Hi @Zethson, huge thanks for taking the time to review the package! It is accurate. I will try to address the missing points and comments as time allows. Here are some explanations, feel free to suggest ways to improve the current state:
>
>     1. The reason we recommend to clone and install the package locally is not because of code dependence, but rather because the tutorial files are included in the repo, but not the package. Do you (or another person reading this thread) have another solution for making the tutorial code available to the user in the simplest way?

I'd move all tutorials into notebooks and add them to a biocypher-tutorials repository. The tutorials themselves can then download data as needed. Ideally, you could also offer colab/binder instances to run these notebooks then. This may require custom containers and will generally require substantial effort. You could also start by adding "download notebook" buttons to the tutorials.

>     3. For me it is. (The first couple functions for setup do not have much of an output.) Could be something to do with how the colab works (each user has their own version, maybe)? I don't usually work with colab, so feel free to suggest improvements.

I'll look again.

>     4. I am using the Sphinx autodoc module for rendering the function documentation. There is already an [issue](https://github.com/biocypher/biocypher/issues/183) for making it work for all modules, but I have not had time to delve deep into Sphinx functionality to figure this out. We do only have a fairly small amount of user-facing functions in the first place, which are explained in the tutorials. I could maybe add a manual explanation of the most important functions in [quickstart](https://github.com/biocypher/biocypher/issues/219), but that would not be automatic any more.

Yeah, you're explaining them in the tutorials, but API documentation for all public functions is imperative IMO.

>     5. You are right, there is no action for CI. I have been doing it manually so far due to lack of time to establish a workflow, but I agree that this takes high priority. One issue is that some tests depend on DBMSs running on the test machine, so this is not a straightforward setup. It is on [the list](https://github.com/biocypher/biocypher/issues/218).

I can totally see how this is a lot of work, but I really feel like it will be time well spent. This will heavily improve maintainability and the ability for other people to contribute.

Thanks!
