### 2023-07-26 #110 @slobentanzer
I have added a separate milestone to track and address the review feedback: https://github.com/biocypher/biocypher/milestone/6
