### 2023-07-27 cce82ca
overhaul quickstart documentation closes #219
