+++ docs/submodule.rst
@@ -1,6 +1,6 @@
-##############################
-Submodule documentation
-##############################
+#############
+API Reference
+#############
 
 ``_core.py``: The main BioCypher interface
 =========================================
