+++ docs/quickstart.md
@@ -29,27 +29,35 @@ If you are new to BioCypher and would like a step-by-step introduction to the
 package, please follow the [tutorial](tutorial).
 ```
 
-The BioCypher workflow of creating your own knowledge graph consists of two
-components:
+The BioCypher workflow of creating your own knowledge graph consists of three
+consecutive steps:
 
-1. the [host module adapter](qs_host-module-adapter), a python
-   program, and
-2. the [schema configuration file](qs_schema-config), a YAML file.
+1. Clearly define the scope of your project, including the data sources you want
+to use, the entities and relationships you want to represent, and the
+[ontologies](ontologies) that should inform these entities.
 
-The adapter serves as a data interface between the source and BioCypher,
-piping the "raw" data into BioCypher for the creation of the property
-graph, while the schema configuration tells BioCypher how the graph
-should be structured, detailing the names of constituents and how they
-should be connected.
+2. Using these definitions, find existing [adapters of data
+sources](qs_input-adapter) or, if necessary, create your own. For the data
+yielded by these adapters, create a [schema configuration](qs_schema-config)
+file that tells BioCypher how to represent the entities and relationships in the
+graph.
 
-(qs_host-module-adapter)=
-## The host module adapter
+3. [Run BioCypher](qs_run-bc) using the adapters and schema config to create the
+knowledge graph. If necessary, iterate over KG construction and configuration
+until you are satisfied with the result.
+
+(qs_input-adapter)=
+## The input adapter
 
 BioCypher follows a modular approach to data inputs; to create a knowledge
 graph, we use at least one adapter module that provides a data stream to build
 the graph from. Examples for current adapters can be found on the [GitHub
 project adapter view](https://github.com/orgs/biocypher/projects/3/views/2).
-Adapters can ingest data from many different input sources, including Python
+This is the first place to look when creating your own KG; BioCypher adapters
+are meant to be reusable and a centralised way of maintaining access to data
+sources.
+
+Adapters can ingest data from many different input formats, including Python
 modules as in the [CROssBAR
 adapter](https://github.com/HUBioDataLab/CROssBAR-BioCypher-Migration) (which
 uses the OmniPath backend software, PyPath, for downloading and caching data),
@@ -57,40 +65,15 @@ advanced file management formats such as Parquet as in the [Open Targets
 adapter](https://github.com/biocypher/open-targets), or simple CSV files as in
 the [Dependency Map adapter](https://github.com/biocypher/dependency-map).
 
-The recommended way of interacting with BioCypher is via the
-{py:class}``biocypher._driver.Driver`` class. It can be called either starting
-in "offline mode" using `offline = True`, i.e., without connection to a running
-Neo4j instance, or by providing authentication details via arguments or
-configuration file:
-
-```{code-block} python
-import biocypher
-d = biocypher.Driver(
-  offline = False,
-  db_uri = "bolt://localhost:7687",
-  db_user = "neo4j",
-  db_passwd = "password",
-)
-```
-
-```{note}
-We use the APOC library for Neo4j, which is not included automatically, but
-needs to be installed as a plugin to the DMBS. For more information, please
-refer to the [APOC documentation](https://neo4j.com/labs/apoc/).
-```
-
-```{hint}
-The settings for the BioCypher driver can also be specified in a configuration
-file. For more details, please refer to the [Setup instructions](config).
-```
-
 The main function of the adapter is to pass data into BioCypher, usually
 as some form of iterable (commonly a list or generator of items). As a
 minimal example, we load a list of proteins with identifiers, trivial
 names, and molecular masses from a (fictional) CSV:
 
 ```{code-block} python
-# read into data frame
+:caption: Adapter yielding nodes
+
+# read data into df
 with open("file.csv", "r") as f:
   proteins = pd.read_csv(f)
 
[…]
