+++ docs/tutorial-ontology.md
@@ -1,3 +1,4 @@
+(ontologies)=
 # Tutorial - Handling Ontologies
 
 ![Adapters and ontologies](figure_modularity.png)
