### 2023-07-27 biocypher/biocypher#217 @slobentanzer
closed by cce82ca388565346a911903f06d6de971cf4437f
