+++ docs/submodule.rst
@@ -1,3 +1,5 @@
+.. _api-reference:
+
 #############
 API Reference
 #############
