### 2023-07-31 096dd26
clean pre-commit config
