+++ .pre-commit-config.yaml
@@ -14,7 +14,6 @@ repos:
     -   id: yapf
         language: python
         additional_dependencies: [toml]
-        # stages: [manual]
 -   repo: https://github.com/Instagram/Fixit
     rev: 9d59f968e84bd2773f34b0069eeeaad3ce783254
     hooks:
@@ -25,19 +24,10 @@ repos:
     hooks:
     -   id: isort
         additional_dependencies: [toml]
-# -   repo: https://github.com/hakancelikdev/pyall
-#     rev: 0.2.0
-#     hooks:
-#     -   id: pyall
-#         args: [--refactor]
 -   repo: https://github.com/snok/pep585-upgrade
     rev: v1.0.1
     hooks:
     -   id: upgrade-type-hints
-# -   repo: https://github.com/asottile/add-trailing-comma
-#     rev: v2.2.3
-#     hooks:
-#     -   id: add-trailing-comma
 -   repo: https://github.com/myint/unify
     rev: v0.5
     hooks:
@@ -60,20 +50,6 @@ repos:
     -   id: fix-encoding-pragma
         args: [--remove] # for Python3 codebase, it's not necessary
     -   id: requirements-txt-fixer
-# -   repo: https://github.com/john-hen/Flake8-pyproject
-#     rev: 1.1.0
-#     hooks:
-#     -   id: Flake8-pyproject
-#         additional_dependencies: [flake8-docstrings, flake8-comprehensions, flake8-bugbear]
-# -   repo: https://github.com/myint/rstcheck
-#     rev: v6.0.0rc3
-#     hooks:
-#     -   id: rstcheck
-# -   repo: https://github.com/asottile/pyupgrade
-#     rev: v2.23.1
-#     hooks:
-#     -   id: pyupgrade
-#         args: [--py3-plus, --py36-plus]
 -   repo: https://github.com/pre-commit/pygrep-hooks
     rev: v1.9.0
     hooks:
