### 2023-07-31 64d77bc
switch to black and remove redundant hooks
