+++ pyproject.toml
@@ -116,14 +116,3 @@ per-file-ignores = [
 ]
 max-line-length = 80
 count = true
-
-[tool.yapf]
-based_on_style = "facebook"
-split_penalty_before_if_expr = 0
-split_penalty_import_names = 0
-split_penalty_comprehension = 0
-split_penalty_for_added_line_split = 0
-split_penalty_after_opening_bracket = 0
-split_before_first_argument = true
-split_before_named_assigns = true
-split_complex_comprehension = true
