+++ .pre-commit-config.yaml
@@ -8,32 +8,21 @@ default_stages:
 -   push
 minimum_pre_commit_version: 2.7.1
 repos:
--   repo: https://github.com/google/yapf
-    rev: v0.32.0
+-   repo: https://github.com/ambv/black
+    rev: 23.7.0
     hooks:
-    -   id: yapf
-        language: python
-        additional_dependencies: [toml]
--   repo: https://github.com/Instagram/Fixit
-    rev: 9d59f968e84bd2773f34b0069eeeaad3ce783254
-    hooks:
-    -   id: fixit-run-rules
-        stages: [manual]
+    -   id: black
 -   repo: https://github.com/timothycrosley/isort
-    rev: 5.10.1
+    rev: 5.12.0
     hooks:
     -   id: isort
         additional_dependencies: [toml]
 -   repo: https://github.com/snok/pep585-upgrade
-    rev: v1.0.1
+    rev: v1.0
     hooks:
     -   id: upgrade-type-hints
--   repo: https://github.com/myint/unify
-    rev: v0.5
-    hooks:
-    -   id: unify
 -   repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v4.2.0
+    rev: v4.4.0
     hooks:
     -   id: check-docstring-first
     -   id: end-of-file-fixer
@@ -51,7 +40,7 @@ repos:
         args: [--remove] # for Python3 codebase, it's not necessary
     -   id: requirements-txt-fixer
 -   repo: https://github.com/pre-commit/pygrep-hooks
-    rev: v1.9.0
+    rev: v1.10.0
     hooks:
     -   id: python-no-eval
     -   id: python-use-type-annotations
