+++ tutorial/05__property_inheritance_pandas.py
@@ -9,11 +9,13 @@ from tutorial.data_generator import (
 def main():
     # Setup: create a list of proteins to be imported
     proteins = [
-        p for sublist in zip(
+        p
+        for sublist in zip(
             [RandomPropertyProtein() for _ in range(10)],
             [RandomPropertyProteinIsoform() for _ in range(10)],
             [EntrezProtein() for _ in range(10)],
-        ) for p in sublist
+        )
+        for p in sublist
     ]
 
     # Extract id, label, and property dictionary
@@ -27,8 +29,8 @@ def main():
 
     # Create BioCypher driver
     bc = BioCypher(
-        biocypher_config_path='tutorial/05_biocypher_config.yaml',
-        schema_config_path='tutorial/05_schema_config.yaml',
+        biocypher_config_path="tutorial/05_biocypher_config.yaml",
+        schema_config_path="tutorial/05_schema_config.yaml",
     )
     # Run the import
     bc.add(node_generator())
@@ -38,5 +40,5 @@ def main():
         print(df)
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     main()
