+++ test/test_write_postgres.py
@@ -4,7 +4,7 @@ import subprocess
 import pytest
 
 
-@pytest.mark.parametrize('l', [4], scope='module')
+@pytest.mark.parametrize("l", [4], scope="module")
 def test_write_node_data_from_gen_comma_postgresql(
     bw_comma_postgresql, _get_nodes
 ):
@@ -20,8 +20,8 @@ def test_write_node_data_from_gen_comma_postgresql(
 
     tmp_path = bw_comma_postgresql.outdir
 
-    p_csv = os.path.join(tmp_path, 'Protein-part000.csv')
-    m_csv = os.path.join(tmp_path, 'MicroRNA-part000.csv')
+    p_csv = os.path.join(tmp_path, "Protein-part000.csv")
+    m_csv = os.path.join(tmp_path, "MicroRNA-part000.csv")
 
     with open(p_csv) as f:
         pr = f.read()
@@ -30,15 +30,15 @@ def test_write_node_data_from_gen_comma_postgresql(
         mi = f.read()
 
     assert 'p1,"StringProperty1",4.0,9606' in pr
-    assert 'uniprot' in pr
-    assert 'BiologicalEntity' in pr
-    assert 'Polypeptide' in pr
-    assert 'Protein' in pr
+    assert "uniprot" in pr
+    assert "BiologicalEntity" in pr
+    assert "Polypeptide" in pr
+    assert "Protein" in pr
     assert 'm1,"StringProperty1",9606,"m1","mirbase"' in mi
-    assert 'ChemicalEntity' in mi
+    assert "ChemicalEntity" in mi
 
 
-@pytest.mark.parametrize('l', [4], scope='module')
+@pytest.mark.parametrize("l", [4], scope="module")
 def test_write_node_data_from_gen_tab_postgresql(bw_tab_postgresql, _get_nodes):
     nodes = _get_nodes
 
@@ -49,8 +49,8 @@ def test_write_node_data_from_gen_tab_postgresql(bw_tab_postgresql, _get_nodes):
 
     tmp_path = bw_tab_postgresql.outdir
 
-    p_csv = os.path.join(tmp_path, 'Protein-part000.csv')
-    m_csv = os.path.join(tmp_path, 'MicroRNA-part000.csv')
+    p_csv = os.path.join(tmp_path, "Protein-part000.csv")
+    m_csv = os.path.join(tmp_path, "MicroRNA-part000.csv")
 
     with open(p_csv) as f:
         pr = f.read()
@@ -61,19 +61,25 @@ def test_write_node_data_from_gen_tab_postgresql(bw_tab_postgresql, _get_nodes):
     assert passed
     assert 'p1\t"StringProperty1"\t4.0\t9606\t' in pr
     assert '\t"uniprot"\t' in pr
-    assert 'BiologicalEntity' in pr
-    assert 'Polypeptide' in pr
-    assert 'Protein' in pr
+    assert "BiologicalEntity" in pr
+    assert "Polypeptide" in pr
+    assert "Protein" in pr
     assert 'm1\t"StringProperty1"\t9606\t"m1"\t"mirbase"' in mi
-    assert 'ChemicalEntity' in mi
+    assert "ChemicalEntity" in mi
 
 
 @pytest.mark.requires_postgresql
-@pytest.mark.parametrize('l', [4], scope='module')
+@pytest.mark.parametrize("l", [4], scope="module")
 def test_database_import_node_data_from_gen_comma_postgresql(
     bw_comma_postgresql, _get_nodes, create_database_postgres
 ):
-    dbname, user, port, password, create_database_success = create_database_postgres
+    (
+        dbname,
+        user,
+        port,
+        password,
+        create_database_success,
+    ) = create_database_postgres
     assert create_database_success
 
[…]
