+++ test/test_integration.py
@@ -3,7 +3,7 @@ import os
 import pytest
 
 
-@pytest.mark.parametrize('l', [4], scope='function')
+@pytest.mark.parametrize("l", [4], scope="function")
 def test_write_node_data_from_gen(core, _get_nodes):
     nodes = _get_nodes
 
@@ -15,8 +15,8 @@ def test_write_node_data_from_gen(core, _get_nodes):
 
     path = core._output_directory
 
-    p_csv = os.path.join(path, 'Protein-part000.csv')
-    m_csv = os.path.join(path, 'MicroRNA-part000.csv')
+    p_csv = os.path.join(path, "Protein-part000.csv")
+    m_csv = os.path.join(path, "MicroRNA-part000.csv")
 
     with open(p_csv) as f:
         pr = f.read()
@@ -26,9 +26,9 @@ def test_write_node_data_from_gen(core, _get_nodes):
 
     assert passed
     assert "p1;'StringProperty1';4.0;9606;'gene1|gene2';'p1';'uniprot'" in pr
-    assert 'BiologicalEntity' in pr
+    assert "BiologicalEntity" in pr
     assert "m1;'StringProperty1';9606;'m1';'mirbase'" in mi
-    assert 'ChemicalEntity' in mi
+    assert "ChemicalEntity" in mi
 
 
 def test_show_ontology_structure_kwargs(core):
