+++ biocypher/__init__.py
@@ -13,14 +13,14 @@ BioCypher: a unifying framework for biomedical knowledge graphs.
 """
 
 __all__ = [
-    '__version__',
-    '__author__',
-    'module_data',
-    'config',
-    'logfile',
-    'log',
-    'Driver',
-    'BioCypher',
+    "__version__",
+    "__author__",
+    "module_data",
+    "config",
+    "logfile",
+    "log",
+    "Driver",
+    "BioCypher",
 ]
 
 from ._core import BioCypher
@@ -30,11 +30,10 @@ from ._metadata import __author__, __version__
 
 
 class Driver(BioCypher):
-
     # initialise parent class but log a warning
     def __init__(self, *args, **kwargs):
         logger.warning(
-            'The class `Driver` is deprecated and will be removed in a future '
-            'release. Please use `BioCypher` instead.'
+            "The class `Driver` is deprecated and will be removed in a future "
+            "release. Please use `BioCypher` instead."
         )
         super().__init__(*args, **kwargs)

[…]
