### 2023-07-31 120c399
move graphical abstract to docs folder
