### 2023-07-31 3afa2c9
remove hardcoded version from `_metadata.py`
