+++ biocypher/_metadata.py
@@ -19,8 +19,6 @@ import importlib.metadata
 
 import toml
 
-_VERSION = "0.5.17"
-
 
 def get_metadata():
     """
@@ -60,7 +58,7 @@ def get_metadata():
         except importlib.metadata.PackageNotFoundError:
             pass
 
-    meta["version"] = meta.get("version", None) or _VERSION
+    meta["version"] = meta.get("version", None)
 
     return meta
 
