+++ .bumpversion.cfg
@@ -1,6 +1,6 @@
 [bumpversion]
 current_version = 0.5.17
-commit = True
+commit = False
 tag = True
 files = pyproject.toml
 parse = (?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)
