+++ .bumpversion.cfg
@@ -1,7 +1,7 @@
 [bumpversion]
 current_version = 0.5.17
 commit = False
-tag = True
+tag = False
 files = pyproject.toml
 parse = (?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)
 serialize = {major}.{minor}.{patch}
