### 2023-07-31 4bd0fbd
return version to `_metadata.py`, test replacement .. of version in there by bumpversion
