+++ .bumpversion.cfg
@@ -2,6 +2,5 @@
 current_version = 0.5.17
 commit = False
 tag = False
-files = [pyproject.toml, biocypher/_metadata.py]
 parse = (?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)
 serialize = {major}.{minor}.{patch}
