+++ .bumpversion.cfg
@@ -4,3 +4,5 @@ commit = False
 tag = False
 parse = (?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)
 serialize = {major}.{minor}.{patch}
+
+[bumpversion:file:pyproject.toml]
