+++ .bumpversion.cfg
@@ -6,3 +6,5 @@ parse = (?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)
 serialize = {major}.{minor}.{patch}
 
 [bumpversion:file:pyproject.toml]
+
+[bumpversion:file:biocypher/_metadata.py]
