+++ .bumpversion.cfg
@@ -1,7 +1,7 @@
 [bumpversion]
 current_version = 0.5.17
-commit = False
-tag = False
+commit = True
+tag = True
 parse = (?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)
 serialize = {major}.{minor}.{patch}
 
