### 2023-07-31 d45d9e7
reactivate autotag; seems more robust
