+++ .bumpversion.cfg
@@ -1,7 +1,7 @@
 [bumpversion]
 current_version = 0.5.18
 commit = True
-tag = False
+tag = True
 parse = (?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)
 serialize = {major}.{minor}.{patch}
 
