### 2023-07-31 d4b6ea0
add `paper.md`
