+++ paper.md
@@ -1,0 +1,121 @@
+---
+title: "Democratizing knowledge representation with BioCypher"
+tags:
+  - Bioinformatics
+  - Knowledge Representation
+  - Computational Biology
+  - Software
+authors:
+  - name: Sebastian Lobentanzer
+    affiliation: [1]
+    corresponding: true
+  - name: Patrick Aloy
+    affiliation: [2, 3]
+  - name: Jan Baumbach
+    affiliation: [4]
+  - name: Balazs Bohar
+    affiliation: [5, 6]
+  - name: Vincent J. Carey
+    affiliation: [7]
+  - name: Pornpimol Charoentong
+    affiliation: [8, 9]
+  - name: Katharina Danhauser
+    affiliation: [10]
+  - name: Tunca Doğan
+    affiliation: [11, 12]
+  - name: Johann Dreo
+    affiliation: [13, 14]
+  - name: Ian Dunham
+    affiliation: [15, 16]
+  - name: Elias Farr
+    affiliation: [1]
+  - name: Adrià Fernandez-Torras
+    affiliation: [2]
+  - name: Benjamin M. Gyori
+    affiliation: [17]
+  - name: Michael Hartung
+    affiliation: [4]
+  - name: Charles Tapley Hoyt
+    affiliation: [17]
+  - name: Christoph Klein
+    affiliation: [10]
+  - name: Tamas Korcsmaros
+    affiliation: [5, 18, 19]
+  - name: Andreas Maier
+    affiliation: [4]
+  - name: Matthias Mann
+    affiliation: [20, 21]
+  - name: David Ochoa
+    affiliation: [15, 16]
+  - name: Elena Pareja-Lorente
+    affiliation: [2]
+  - name: Ferdinand Popp
+    affiliation: [22]
+  - name: Martin Preusse
+    affiliation: [23]
+  - name: Niklas Probul
+    affiliation: [4]
+  - name: Benno Schwikowski
+    affiliation: [13]
+  - name: Bünyamin Sen
+    affiliation: [11, 12]
+  - name: Maximilian T. Strauss
+    affiliation: [20]
+  - name: Denes Turei
+    affiliation: [1]
+  - name: Erva Ulusoy
+    affiliation: [11, 12]
+  - name: Dagmar Waltemath
+    affiliation: [24]
+  - name: Judith A. H. Wodke
+    affiliation: [24]
+  - name: Julio Saez-Rodriguez
+    affiliation: [1]
+    corresponding: true
+affiliations:
+  - name: Heidelberg University, Faculty of Medicine, and Heidelberg University Hospital, Institute for Computational Biomedicine, Bioquant
+    index: 1
+  - name: Institute for Research in Biomedicine (IRB Barcelona), the Barcelona Institute of Science and Technology
+    index: 2
+  - name: Institució Catalana de Recerca i Estudis Avançats (ICREA)
+    index: 3
+  - name: Institute for Computational Systems Biology, University of Hamburg, Germany
+    index: 4
+  - name: Earlham Institute, Norwich, UK
+    index: 5
+  - name: Biological Research Centre, Szeged, Hungary
+    index: 6
+  - name: Channing Division of Network Medicine, Mass General Brigham, Harvard Medical School, Boston, USA
+    index: 7
+  - name: Centre for Quantitative Analysis of Molecular and Cellular Biosystems (Bioquant), Heidelberg University
+    index: 8
+  - name: Department of Medical Oncology, National Centre for Tumour Diseases (NCT), Heidelberg University Hospital (UKHD)
+    index: 9
+  - name: Department of Pediatrics, Dr. von Hauner Children’s Hospital, University Hospital, LMU Munich, Germany
+    index: 10
[…]
