### 2023-08-01 d30d5f5
add paper.md sections; closes #222
