### 2023-08-01 cc3be57
add CI/CD workflow that depends on docker .. add --host param to postgres to allow conection
