+++ test/test_write_postgres.py
@@ -121,7 +121,7 @@ def test_database_import_node_data_from_gen_comma_postgresql(
         assert result.returncode == 0
 
     # check data in the databases
-    command = f"PGPASSWORD={password} psql -c 'SELECT COUNT(*) FROM protein;' --dbname {dbname} --port {port} --user {user}"
+    command = f"PGPASSWORD={password} psql -c 'SELECT COUNT(*) FROM protein;' --dbname {dbname} --host localhost --port {port} --user {user}"
     result = subprocess.run(
         command, shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT
     )
@@ -131,7 +131,7 @@ def test_database_import_node_data_from_gen_comma_postgresql(
     assert "4" in result.stdout.decode()
 
     # check data in the databases
-    command = f"PGPASSWORD={password} psql -c 'SELECT COUNT(*) FROM microrna;' --dbname {dbname} --port {port} --user {user}"
+    command = f"PGPASSWORD={password} psql -c 'SELECT COUNT(*) FROM microrna;' --dbname {dbname} --host localhost --port {port} --user {user}"
     result = subprocess.run(
         command, shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT
     )
@@ -194,7 +194,7 @@ def test_database_import_node_data_from_gen_tab_postgresql(
         assert result.returncode == 0
 
     # check data in the databases
-    command = f"PGPASSWORD={password} psql -c 'SELECT COUNT(*) FROM protein;' --dbname {dbname} --port {port} --user {user}"
+    command = f"PGPASSWORD={password} psql -c 'SELECT COUNT(*) FROM protein;' --dbname {dbname} --host localhost --port {port} --user {user}"
     result = subprocess.run(
         command, shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT
     )
@@ -204,7 +204,7 @@ def test_database_import_node_data_from_gen_tab_postgresql(
     assert "5" in result.stdout.decode()
 
     # check data in the databases
-    command = f"PGPASSWORD={password} psql -c 'SELECT COUNT(*) FROM microrna;' --dbname {dbname} --port {port} --user {user}"
+    command = f"PGPASSWORD={password} psql -c 'SELECT COUNT(*) FROM microrna;' --dbname {dbname} --host localhost --port {port} --user {user}"
     result = subprocess.run(
         command, shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT
     )
@@ -272,7 +272,7 @@ def test_database_import_edge_data_from_gen_comma_postgresql(
     assert result.returncode == 0
 
     # check data in the databases
-    command = f"PGPASSWORD={password} psql -c 'SELECT COUNT(*) FROM is_mutated_in;' --dbname {dbname} --port {port} --user {user}"
+    command = f"PGPASSWORD={password} psql -c 'SELECT COUNT(*) FROM is_mutated_in;' --dbname {dbname} --host localhost --port {port} --user {user}"
     result = subprocess.run(
         command, shell=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT
     )
@@ -281,7 +281,7 @@ def test_database_import_edge_data_from_gen_comma_postgresql(
     # 2 entries in table
     assert "8" in result.stdout.decode()
 
-    command = f"PGPASSWORD={password} psql -c 'SELECT COUNT(*) FROM perturbed_in_disease;' --dbname {dbname} --port {port} --user {user}"
[…]
