+++ .github/workflows/ci_cd.yaml
@@ -31,8 +31,11 @@ jobs:
     - name: Start Neo4j Docker
       run: docker run --restart always --publish=7474:7474 --publish=7687:7687 --env NEO4J_AUTH=neo4j/your_password_here --env NEO4J_PLUGINS='["apoc"]' --env=NEO4J_ACCEPT_LICENSE_AGREEMENT=yes -d neo4j:4.4-enterprise
 
+    - name: Start Postgres Docker
+      run: docker run --name pg --restart always --publish=5432:5432 --env POSTGRES_PASSWORD=postgres -d postgres:latest
+
     - name: Run Tests
-      run: coverage run -m pytest
+      run: coverage run -m pytest --password=your_password_here
 
     - name: Generate coverage badge
       run: coverage-badge -f -o docs/coverage.svg
