### 2023-08-01 b0630c5
add path assert
