+++ pyproject.toml
@@ -54,6 +54,7 @@ isort = "^5.10.1"
 ipython = "^8.7.0"
 ipykernel = "^6.23.1"
 sphinxext-opengraph = "^0.8.2"
+coverage-badge = "^1.1.0"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
