### 2023-08-01 1fdb2d4
add CI and coverage badges, links, yml -> yaml
