### 2023-08-01 biocypher/biocypher#218 @slobentanzer
Closed by 2a3c887d96b3c06faced83d817a050c9b8843a5d
