### 2023-08-01 8d256cb
switch to autosummary for API reference, .. refactor documentation, clean up links and outdated information, .. rename submodule documentation to API
