+++ docs/submodule.rst
@@ -1,50 +1,0 @@
-.. _api-reference:
-
-#############
-API Reference
-#############
-
-``_core.py``: The main BioCypher interface
-=========================================
-.. automodule:: biocypher._core
-   :members:
-
-``_write.py``: Write the Graph to various formats for batch import
-==================================================================
-.. automodule:: biocypher._write
-   :members:
-
-``_connect.py``: On-line functionality for interaction with a DBMS
-===============================================================
-.. automodule:: biocypher._connect
-   :members:
-
-``_mapping.py``: Mapping of data inputs to KG ontology
-===================================================
-.. automodule:: biocypher._mapping
-   :members:
-
-``_ontology.py``: Ontology ingestion, parsing, and manipulation
-==============================================================
-.. automodule:: biocypher._ontology
-   :members:
-
-``_create.py``: Base classes for node and edge representations in BioCypher
-=========================================================================
-.. automodule:: biocypher._create
-   :members:
-
-``_translate.py``: Translation functionality for implemented types of representation
-==================================================================================
-.. automodule:: biocypher._translate
-   :members:
-
-``_logger.py``: Logging
-=====================
-.. automodule:: biocypher._logger
-   :members:
-
-``_misc.py``: Miscellaneous utility functions
-===========================================
-.. automodule:: biocypher._misc
-   :members:
