+++ docs/tutorial-ontology.md
@@ -1,4 +1,4 @@
-(ontologies)=
+(tutorial_ontologies)=
 # Tutorial - Handling Ontologies
 
 ![Adapters and ontologies](figure_modularity.png)

[…]
