+++ _sources/modules/biocypher._misc.create_tree_visualisation.rst.txt
@@ -1,0 +1,6 @@
+﻿biocypher.\_misc.create\_tree\_visualisation
+============================================
+
+.. currentmodule:: biocypher._misc
+
+.. autofunction:: create_tree_visualisation
