+++ _sources/modules/biocypher._misc.snakecase_to_sentencecase.rst.txt
@@ -1,0 +1,6 @@
+﻿biocypher.\_misc.snakecase\_to\_sentencecase
+============================================
+
+.. currentmodule:: biocypher._misc
+
+.. autofunction:: snakecase_to_sentencecase
