### 2023-08-01 164ab3d
remove pandas tutorials, rename notebook
