### 2023-08-01 e6d61fb
render jupyter notebook as HTML tutorial using .. nbsphinx; fix links, move notebook to docs folder
