+++ docs/contents.rst
@@ -8,6 +8,7 @@
    installation
    quickstart
    adapters
+   notebooks/pandas_tutorial.ipynb
    tutorial
    tutorial-ontology
    tutorial-adapter
