+++ pyproject.toml
@@ -55,6 +55,7 @@ ipython = "^8.7.0"
 ipykernel = "^6.23.1"
 sphinxext-opengraph = "^0.8.2"
 coverage-badge = "^1.1.0"
+nbsphinx = "^0.9.2"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
