### 2023-08-01 f73b3d3
install pandoc on workflow runner
