+++ .github/workflows/sphinx_autodoc.yaml
@@ -36,6 +36,8 @@ jobs:
             run: poetry install --no-interaction --no-root
         -   name: Install library
             run: poetry install --no-interaction
+        -   name: Install pandoc
+            run: sudo apt-get -y install pandoc
         -   name: Build documentation
             run: poetry run make html --directory docs/
         -   name: Commit files
