### 2023-08-01 ffaf4c4
try different coverage strategy (publish subdir)
