+++ .github/workflows/ci_cd.yaml
@@ -45,5 +45,5 @@ jobs:
       env:
         REPO: self
         BRANCH: coverage
-        FOLDER: coverage
+        FOLDER: docs/coverage
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
