+++ biocypher/_get.py
@@ -12,6 +12,8 @@
 BioCypher get module. Used to download and cache data from external sources.
 """
 
+from __future__ import annotations
+
 from ._logger import logger
 
 logger.debug(f"Loading module {__name__}.")
@@ -20,12 +22,19 @@ from datetime import datetime, timedelta
 from tempfile import TemporaryDirectory
 import os
 import json
+import ftplib
 
 import pooch
 
 
 class Resource:
-    def __init__(self, name: str, url: str, lifetime: int = 0):
+    def __init__(
+        self,
+        name: str,
+        url_s: str | list[str],
+        lifetime: int = 0,
+        is_dir: bool = False,
+    ):
         """
         A resource is a file that can be downloaded from a URL and cached
         locally. This class implements checks of the minimum requirements for
@@ -40,8 +49,9 @@ class Resource:
                 resource is considered to be permanent.
         """
         self.name = name
-        self.url = url
+        self.url_s = url_s
         self.lifetime = lifetime
+        self.is_dir = is_dir
 
 
 class Downloader:
@@ -74,8 +84,9 @@ class Downloader:
         for resource in resources:
             paths.append(self._download_or_cache(resource))
 
-        # flatten list
-        paths = [path for sublist in paths for path in sublist]
+        # flatten list if it is nested
+        if is_nested(paths):
+            paths = [path for sublist in paths for path in sublist]
 
         return paths
 
@@ -104,14 +115,29 @@ class Downloader:
         if expired or not cache:
             logger.info(f"Downloading resource {resource.name}.")
 
-            path = pooch.retrieve(
-                resource.url,
-                None,
-                fname=resource.name,
-                path=self.cache_dir,
-                processor=pooch.Unzip(),
-                progressbar=True,
-            )
+            if resource.is_dir:
+                files = self._get_files(resource)
+                resource.url_s = [resource.url_s + "/" + file for file in files]
+                resource.is_dir = False
+                paths = self._download_or_cache(resource, cache)
+            elif isinstance(resource.url_s, list):
+                paths = []
+                for url in resource.url_s:
+                    fname = resource.name + "_" + url[url.rfind("/") + 1 :]
+                    paths.append(
+                        self._retrieve(
+                            url=url,
+                            fname=fname,
+                            path=os.path.join(self.cache_dir, resource.name),
+                        )
+                    )
+            else:
+                fname = resource.url_s[resource.url_s.rfind("/") + 1 :]
+                paths = self._retrieve(
+                    url=resource.url_s,
+                    fname=fname,
+                    path=os.path.join(self.cache_dir, resource.name),
+                )
 
             # sometimes a compressed file contains multiple files
             # TODO ask for a list of files in the archive to be used from the
@@ -120,7 +146,90 @@ class Downloader:
             # update cache record
             self._update_cache_record(resource)
 
-            return path
+            return paths
+
+    def _retrieve(
+        self,
+        url: str,
+        fname: str,
+        path: str,
+        known_hash: str = None,
+    ):
+        """
+        Retrieve a file from a URL using Pooch. Infer type of file from
+        extension and use appropriate processor.
+
+        Args:
+            url (str): The URL to retrieve the file from.
+
+            fname (str): The name of the file.
+
+            path (str): The path to the file.
+        """
+        if fname.endswith(".zip"):
+            return pooch.retrieve(
+                url=url,
+                known_hash=known_hash,
+                fname=fname,
+                path=path,
+                processor=pooch.Unzip(),
+                progressbar=True,
+            )
+
+        elif fname.endswith(".tar.gz"):
+            return pooch.retrieve(
+                url=url,
+                known_hash=known_hash,
+                fname=fname,
[…]
