### 2023-08-07 #110 @slobentanzer
Hi @arianesasso,
as mentioned, you can see the progress of the revision in the milestone (most things from @Zethson are addressed already). I am waiting for the second review before doing more.
