### 2023-08-15 #110 @Midnighter
Hi everyone,

Just to update you, that we started going through the tutorials and looking at the codebase, but won't be able to complete the review today. Our plan is to do so on Friday.
