### 2023-08-22 #110 @arianesasso
Thank you @Midnighter and Sinna for the great review! 😄

@slobentanzer would two weeks be a good timeline for you to address the reviews, or is it too tight? Also, let us know in case you have any general or specific questions about them :).
