### 2023-08-23 #110 @slobentanzer
@Midnighter and Sinna, thanks a lot for your time and thoughtful comments. I agree with your review, and have added corresponding issues to the [review milestone](https://github.com/biocypher/biocypher/milestone/6).

@arianesasso the coming weeks are pretty full of deadlines, I am not sure I will be able to address all points in 2 weeks. End of September seems more realistic, but I will update here in any case (plus any questions that may arise).
