### 2023-08-23 fd559c1
core getter function
