+++ biocypher/_write.py
@@ -125,7 +125,6 @@ class _BatchWriter(ABC):
 
     def __init__(
         self,
-        ontology: "Ontology",
         translator: "Translator",
         deduplicator: "Deduplicator",
         delimiter: str,
@@ -167,10 +166,6 @@ class _BatchWriter(ABC):
             - _get_import_script_name
 
         Args:
-            ontology:
-                Instance of :py:class:`Ontology` to enable translation and
-                ontology queries
-
             translator:
                 Instance of :py:class:`Translator` to enable translation of
                 nodes and manipulation of properties.
@@ -251,8 +246,6 @@ class _BatchWriter(ABC):
         self.wipe = wipe
         self.strict_mode = strict_mode
 
-        self.extended_schema = ontology.extended_schema
-        self.ontology = ontology
         self.translator = translator
         self.deduplicator = deduplicator
         self.node_property_dict = {}
@@ -451,8 +444,12 @@ class _BatchWriter(ABC):
                     bin_l[label] = 1
 
                     # get properties from config if present
-                    cprops = self.extended_schema.get(label).get(
-                        "properties",
+                    cprops = (
+                        self.translator.ontology.mapping.extended_schema.get(
+                            label
+                        ).get(
+                            "properties",
+                        )
                     )
                     if cprops:
                         d = dict(cprops)
@@ -486,7 +483,7 @@ class _BatchWriter(ABC):
 
                     # get label hierarchy
                     # multiple labels:
-                    all_labels = self.ontology.get_ancestors(label)
+                    all_labels = self.translator.ontology.get_ancestors(label)
 
                     if all_labels:
                         # convert to pascal case
@@ -706,13 +703,23 @@ class _BatchWriter(ABC):
                     # (may not be if it is an edge that carries the
                     # "label_as_edge" property)
                     cprops = None
-                    if label in self.extended_schema:
-                        cprops = self.extended_schema.get(label).get(
+                    if (
+                        label
+                        in self.translator.ontology.mapping.extended_schema
+                    ):
+                        cprops = self.translator.ontology.mapping.extended_schema.get(
+                            label
+                        ).get(
                             "properties",
                         )
                     else:
                         # try via "label_as_edge"
-                        for k, v in self.extended_schema.items():
+                        for (
+                            k,
+                            v,
+                        ) in (
+                            self.translator.ontology.mapping.extended_schema.items()
+                        ):
                             if isinstance(v, dict):
                                 if v.get("label_as_edge") == label:
                                     cprops = v.get("properties")
@@ -873,9 +880,14 @@ class _BatchWriter(ABC):
 
             if label in ["IS_SOURCE_OF", "IS_TARGET_OF", "IS_PART_OF"]:
                 skip_id = True
-            elif not self.extended_schema.get(label):
+            elif not self.translator.ontology.mapping.extended_schema.get(
+                label
+            ):
                 # find label in schema by label_as_edge
-                for k, v in self.extended_schema.items():
+                for (
+                    k,
+                    v,
+                ) in self.translator.ontology.mapping.extended_schema.items():
                     if v.get("label_as_edge") == label:
                         schema_label = k
                         break
@@ -884,7 +896,9 @@ class _BatchWriter(ABC):
 
             if schema_label:
                 if (
-                    self.extended_schema.get(schema_label).get("use_id")
+                    self.translator.ontology.mapping.extended_schema.get(
+                        schema_label
+                    ).get("use_id")
                     == False
                 ):
                     skip_id = True
@@ -1182,9 +1196,14 @@ class _Neo4jBatchWriter(_BatchWriter):
 
             if label in ["IS_SOURCE_OF", "IS_TARGET_OF", "IS_PART_OF"]:
                 skip_id = True
[…]
