+++ test/test_core.py
@@ -29,6 +29,16 @@ def test_log_duplicates(core, deduplicator, _get_nodes):
     assert True
 
 
+@pytest.mark.parametrize("l", [4], scope="module")
+def test_write_schema_info(core, _get_nodes, _get_edges):
+    core.add(_get_nodes)
+    core.add(_get_edges)
+
+    # create rel as node situation
+    relasnodes = None
+    assert core.write_schema_info() == None
+
+
 # def test_access_translate(driver):
 
 #     driver.start_ontology()

[…]
