### 2023-08-24 196c941
adjust test fixtures to refactor
