### 2023-08-24 a3ded24
more conftest fixes
