### 2023-08-24 34cc67b
remove ontology from more places
