### 2023-08-24 fef2c9b
fix pandas init
