### 2023-08-24 44baa16
rel as node functionality in pandas closes #267
