### 2023-08-24 663cc14
move edge deduplication upstream into the .. `write_edges` function to account for RelAsNode relationships introduces a difference between `write_nodes` and `write_edges`, as the .. former still checks in the downstream write function. this is .. preferable because it allows to stay in generator mode for longer
