### 2023-08-24 b1a6211
return and write an extended schema that contains .. - information about presence in the KG - information on the entity being a relationship
