+++ test/test_core.py
@@ -1,3 +1,6 @@
+import os
+
+import yaml
 import pytest
 
 
@@ -35,7 +38,50 @@ def test_write_schema_info(core, _get_nodes, _get_edges, _get_rel_as_nodes):
     core.add(_get_edges)
     core.add(_get_rel_as_nodes)
 
-    assert core.write_schema_info() == None
+    schema = core.write_schema_info()
+
+    assert schema.get("is_schema_info") == True
+    assert schema.get("protein").get("present_in_knowledge_graph") == True
+    assert schema.get("protein").get("is_relationship") == False
+    assert schema.get("microRNA").get("present_in_knowledge_graph") == True
+    assert schema.get("microRNA").get("is_relationship") == False
+    assert (
+        schema.get("gene to disease association").get(
+            "present_in_knowledge_graph"
+        )
+        == True
+    )
+    assert (
+        schema.get("gene to disease association").get("is_relationship") == True
+    )
+    assert (
+        schema.get("mutation to tissue association").get(
+            "present_in_knowledge_graph"
+        )
+        == True
+    )
+    assert (
+        schema.get("mutation to tissue association").get("is_relationship")
+        == True
+    )
+    assert (
+        schema.get("post translational interaction").get(
+            "present_in_knowledge_graph"
+        )
+        == True
+    )
+    assert (
+        schema.get("post translational interaction").get("is_relationship")
+        == True
+    )
+
+    path = os.path.join(core._output_directory, "schema_info.yaml")
+    assert os.path.exists(path)
+
+    with open(path, "r") as f:
+        schema_loaded = yaml.safe_load(f)
+
+    assert schema_loaded == schema
 
 
 # def test_access_translate(driver):
