### 2023-08-24 e80b34c
account for non-dict entities
