+++ test/test_core.py
@@ -21,7 +21,7 @@ def test_log_missing_types(core, translator):
     assert mt.get("a") == 1 and mt.get("b") == 2
 
 
-@pytest.mark.parametrize("l", [4], scope="module")
+@pytest.mark.parametrize("l", [4], scope="function")
 def test_log_duplicates(core, deduplicator, _get_nodes):
     core._deduplicator = deduplicator
     nodes = _get_nodes + _get_nodes
@@ -29,10 +29,13 @@ def test_log_duplicates(core, deduplicator, _get_nodes):
     core.add(nodes)
     core.log_duplicates()
 
-    assert True
+    assert "protein" in core._deduplicator.duplicate_entity_types
+    assert "p1" in core._deduplicator.duplicate_entity_ids
+    assert "microRNA" in core._deduplicator.duplicate_entity_types
+    assert "m1" in core._deduplicator.duplicate_entity_ids
 
 
-@pytest.mark.parametrize("l", [4], scope="module")
+@pytest.mark.parametrize("l", [4], scope="function")
 def test_write_schema_info(core, _get_nodes, _get_edges, _get_rel_as_nodes):
     core.add(_get_nodes)
     core.add(_get_edges)
