+++ test/conftest.py
@@ -1,5 +1,6 @@
 import os
 import random
+import shutil
 import string
 import tempfile
 import subprocess
@@ -51,10 +52,34 @@ def pytest_addoption(parser):
         )
 
 
-# temporary output paths
-def get_random_string(length):
-    letters = string.ascii_lowercase
-    return "".join(random.choice(letters) for _ in range(length))
+@pytest.fixture(scope="session")
+def tmp_path_session(tmp_path_factory):
+    """
+    Create a session-scoped temporary directory.
+
+    Args:
+        tmp_path_factory: The built-in pytest fixture.
+
+    Returns:
+        pathlib.Path: The path to the temporary directory.
+    """
+    return tmp_path_factory.mktemp("data")
+
+
+@pytest.fixture(scope="session", autouse=True)
+def cleanup(request, tmp_path_session):
+    """
+    Teardown function to delete the session-scoped temporary directory.
+
+    Args:
+        request: The pytest request object.
+        tmp_path_session: The session-scoped temporary directory.
+    """
+
+    def remove_tmp_dir():
+        shutil.rmtree(tmp_path_session)
+
+    request.addfinalizer(remove_tmp_dir)
 
 
 # biocypher node generator
@@ -228,11 +253,11 @@ def mondo_adapter():
 
 # neo4j batch writer fixtures
 @pytest.fixture(scope="function")
-def bw(translator, deduplicator, tmp_path):
+def bw(translator, deduplicator, tmp_path_session):
     bw = _Neo4jBatchWriter(
         translator=translator,
         deduplicator=deduplicator,
-        output_directory=tmp_path,
+        output_directory=tmp_path_session,
         delimiter=";",
         array_delimiter="|",
         quote="'",
@@ -241,18 +266,17 @@ def bw(translator, deduplicator, tmp_path):
     yield bw
 
     # teardown
-    for f in os.listdir(tmp_path):
-        os.remove(os.path.join(tmp_path, f))
-    os.rmdir(tmp_path)
+    for f in os.listdir(tmp_path_session):
+        os.remove(os.path.join(tmp_path_session, f))
 
 
 # neo4j batch writer fixtures
 @pytest.fixture(scope="function")
-def bw_tab(translator, deduplicator, tmp_path):
+def bw_tab(translator, deduplicator, tmp_path_session):
     bw_tab = _Neo4jBatchWriter(
         translator=translator,
         deduplicator=deduplicator,
-        output_directory=tmp_path,
+        output_directory=tmp_path_session,
         delimiter="\\t",
         array_delimiter="|",
         quote="'",
@@ -261,17 +285,16 @@ def bw_tab(translator, deduplicator, tmp_path):
     yield bw_tab
 
     # teardown
-    for f in os.listdir(tmp_path):
-        os.remove(os.path.join(tmp_path, f))
-    os.rmdir(tmp_path)
+    for f in os.listdir(tmp_path_session):
+        os.remove(os.path.join(tmp_path_session, f))
 
 
 @pytest.fixture(scope="function")
-def bw_strict(translator, deduplicator, tmp_path):
+def bw_strict(translator, deduplicator, tmp_path_session):
     bw = _Neo4jBatchWriter(
         translator=translator,
         deduplicator=deduplicator,
-        output_directory=tmp_path,
+        output_directory=tmp_path_session,
         delimiter=";",
         array_delimiter="|",
         quote="'",
@@ -281,14 +304,13 @@ def bw_strict(translator, deduplicator, tmp_path):
     yield bw
 
     # teardown
-    for f in os.listdir(tmp_path):
-        os.remove(os.path.join(tmp_path, f))
-    os.rmdir(tmp_path)
+    for f in os.listdir(tmp_path_session):
[…]
