### 2023-08-24 a023d8a
string formatting and unused imports
