### 2023-08-24 a6721d3
expand test and make all URLs lists (also .. single entries)
