### 2023-08-24 43442aa
extend CI/CD to mac and windows closes #195
