### 2023-08-24 81f4fa9
don't prepend resource name, this is taken care .. of by the subdirectory
