+++ .github/workflows/ci_cd.yaml
@@ -19,11 +19,11 @@ jobs:
       with:
         python-version: '3.9'
 
-    - name: Install Poetry
-      run: |
-        curl -sSL https://install.python-poetry.org | python -
-        poetry --version
-        poetry config virtualenvs.create false
+    - name: Install and configure Poetry
+      uses: snok/install-poetry@v1
+      with:
+        version: 1.5.1
+        virtualenvs-create: false
 
     - name: Install Dependencies
       run: poetry install
@@ -45,6 +45,8 @@ jobs:
       env:
         REPO: self
         BRANCH: coverage
+        FOLDER: docs/coverage
+        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
 
   build-macos:
     runs-on: macos-latest
@@ -58,11 +60,11 @@ jobs:
       with:
         python-version: '3.9'
 
-    - name: Install Poetry
-      run: |
-        curl -sSL https://install.python-poetry.org | python -
-        poetry --version
-        poetry config virtualenvs.create false
+    - name: Install and configure Poetry
+      uses: snok/install-poetry@v1
+      with:
+        version: 1.5.1
+        virtualenvs-create: false
 
     - name: Install Dependencies
       run: poetry install
@@ -88,11 +90,11 @@ jobs:
       with:
         python-version: '3.9'
 
-    - name: Install Poetry
-      run: |
-        (Invoke-WebRequest -Uri https://raw.githubusercontent.com/python-poetry/poetry/master/install-poetry.py -UseBasicParsing).Content | python -
-        poetry --version
-        poetry config virtualenvs.create false
+    - name: Install and configure Poetry
+      uses: snok/install-poetry@v1
+      with:
+        version: 1.5.1
+        virtualenvs-create: false
 
     - name: Install Dependencies
       run: poetry install
