+++ .github/workflows/ci_cd.yaml
@@ -69,6 +69,9 @@ jobs:
     - name: Install Dependencies
       run: poetry install
 
+    - name: Install Docker
+      uses: douglascamata/setup-docker-macos-action@v1-alpha
+
     - name: Start Neo4j Docker
       run: docker run --restart always --publish=7474:7474 --publish=7687:7687 --env NEO4J_AUTH=neo4j/your_password_here --env NEO4J_PLUGINS='["apoc"]' --env=NEO4J_ACCEPT_LICENSE_AGREEMENT=yes -d neo4j:4.4-enterprise
 
