### 2023-08-24 50e30a1
set windows runner shell to bash
