+++ .github/workflows/ci_cd.yaml
@@ -84,6 +84,10 @@ jobs:
   build-windows:
     runs-on: windows-latest
 
+    defaults:
+      run:
+        shell: bash
+
     steps:
     - name: Checkout Repository
       uses: actions/checkout@v2
