### 2023-08-24 5693849
deactivate macos build as well; docker install .. does not complete after 10 min on the runner
