+++ .github/workflows/ci_cd.yaml
@@ -48,38 +48,38 @@ jobs:
         FOLDER: docs/coverage
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
 
-  build-macos:
-    runs-on: macos-latest
+  # build-macos:
+  #   runs-on: macos-latest
 
-    steps:
-    - name: Checkout Repository
-      uses: actions/checkout@v2
+  #   steps:
+  #   - name: Checkout Repository
+  #     uses: actions/checkout@v2
 
-    - name: Set up Python
-      uses: actions/setup-python@v2
-      with:
-        python-version: '3.9'
+  #   - name: Set up Python
+  #     uses: actions/setup-python@v2
+  #     with:
+  #       python-version: '3.9'
 
-    - name: Install and configure Poetry
-      uses: snok/install-poetry@v1
-      with:
-        version: 1.5.1
-        virtualenvs-create: false
+  #   - name: Install and configure Poetry
+  #     uses: snok/install-poetry@v1
+  #     with:
+  #       version: 1.5.1
+  #       virtualenvs-create: false
 
-    - name: Install Dependencies
-      run: poetry install
+  #   - name: Install Dependencies
+  #     run: poetry install
 
-    - name: Install Docker
-      uses: douglascamata/setup-docker-macos-action@v1-alpha
+  #   - name: Install Docker
+  #     uses: douglascamata/setup-docker-macos-action@v1-alpha
 
-    - name: Start Neo4j Docker
-      run: docker run --restart always --publish=7474:7474 --publish=7687:7687 --env NEO4J_AUTH=neo4j/your_password_here --env NEO4J_PLUGINS='["apoc"]' --env=NEO4J_ACCEPT_LICENSE_AGREEMENT=yes -d neo4j:4.4-enterprise
+  #   - name: Start Neo4j Docker
+  #     run: docker run --restart always --publish=7474:7474 --publish=7687:7687 --env NEO4J_AUTH=neo4j/your_password_here --env NEO4J_PLUGINS='["apoc"]' --env=NEO4J_ACCEPT_LICENSE_AGREEMENT=yes -d neo4j:4.4-enterprise
 
-    - name: Start Postgres Docker
-      run: docker run --restart always --publish=5432:5432 --env POSTGRES_PASSWORD=postgres -d postgres:latest
+  #   - name: Start Postgres Docker
+  #     run: docker run --restart always --publish=5432:5432 --env POSTGRES_PASSWORD=postgres -d postgres:latest
 
-    - name: Run Tests
-      run: pytest --password=your_password_here
+  #   - name: Run Tests
+  #     run: pytest --password=your_password_here
 
   # build-windows:
   #   runs-on: windows-latest
