### 2023-08-24 bcc88ce
more pandas docs
