### 2023-08-24 3811a4b
make download available from core, cache .. directory setting, get module documentation
