+++ pyproject.toml
@@ -28,7 +28,7 @@ python = "^3.9"
 PyYAML = ">=5.0"
 more_itertools = "*"
 appdirs = "*"
-treelib = "^1.6.1"
+treelib = "1.6.4"
 rdflib = "^6.2.0"
 networkx = "^3.0"
 stringcase = "^1.2.0"
