### 2023-09-12 5f9bb63
fix Issue #250: format badges as table
