### 2023-09-12 d70dc11
fix issue #249: clarify needed docstring format in DEVELOPER.md and some general improvements
