### 2023-09-12 737aca9
fix issue #249: md formatting
