### 2023-09-12 3512535
fix issue #248: include infos about dependency groups in DEVELOPER.md
