### 2023-09-12 b3df89e
fix issue #246: add CITATION.cff
