### 2023-09-12 1318a7c
fix issue #258: update checkout and python-setup actions in CI pipeline
