+++ .github/workflows/ci_cd.yaml
@@ -12,10 +12,10 @@ jobs:
 
     steps:
     - name: Checkout Repository
-      uses: actions/checkout@v2
+      uses: actions/checkout@v4
 
     - name: Set up Python
-      uses: actions/setup-python@v2
+      uses: actions/setup-python@v4
       with:
         python-version: '3.9'
 
@@ -53,10 +53,10 @@ jobs:
 
   #   steps:
   #   - name: Checkout Repository
-  #     uses: actions/checkout@v2
+  #     uses: actions/checkout@v4
 
   #   - name: Set up Python
-  #     uses: actions/setup-python@v2
+  #     uses: actions/setup-python@v4
   #     with:
   #       python-version: '3.9'
 
@@ -90,10 +90,10 @@ jobs:
 
   #   steps:
   #   - name: Checkout Repository
-  #     uses: actions/checkout@v2
+  #     uses: actions/checkout@v4
 
   #   - name: Set up Python
-  #     uses: actions/setup-python@v2
+  #     uses: actions/setup-python@v4
   #     with:
   #       python-version: '3.9'
 
