### 2023-09-12 68f73a0
start to fix issue #252: first code snippet in docs executed to identify outdated function calls
