### 2023-09-13 2038b91
fix issue #252: execute code snippets in documentation to identify outdated function calls
