### 2023-09-13 eb6ffa8
issue #259: run tests and code quality check on every commit
