### 2023-09-13 1a76aaf
issue #259 first draft of release cicd pipeline
