+++ .github/workflows/ci_cd.yaml
@@ -1,4 +1,4 @@
-name: Tests
+name: Tests and code quality
 
 on:
   pull_request:
