### 2023-09-13 0234239
issue #259 fix yaml structure
