+++ .github/workflows/release.yaml
@@ -65,7 +65,6 @@ jobs:
                 FOLDER: docs/_build/html
                 GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
 
-jobs:
     build_and_deploy_artifact:
     steps:
     - name: Checkout Repository
