### 2023-09-13 ad7b087
issue #259 add runner config
