### 2023-09-15 9534a5b
feat(neo4j): supports Neo4j 5+
Neo4j 5.0 refactored the commands of neo4j-admin, which is used by Biocypher.
It introduced backward-incompatible changes:
- commands are now "scoped" with an additional keyword,
- the import command needs an additional `full` keyword,
- the `--database` option is removed in favor of a positional argument,
- the `--force` option is replaced by `--overwrite-destination`.

This patch allows to detect which version of Neo4j is available, and changes
the command configuration accordingly.

As a side effect, existing tests had to be a little bit downgraded.
