+++ biocypher/_write.py
@@ -15,6 +15,7 @@ suitable for import into a DBMS.
 
 import re
 import glob
+import subprocess # neo4j version detection.
 
 from ._logger import logger
 
@@ -1044,6 +1045,31 @@ class _Neo4jBatchWriter(_BatchWriter):
         - _write_array_string
     """
 
+    def __init__(self, *args, **kwargs):
+        """
+        Constructor.
+
+        Check the version of Neo4j and adds a command scope if version >= 5.
+
+        Returns:
+            _Neo4jBatchWriter: An instance of the writer.
+        """
+
+        # Should read the configuration and setup import_call_bin_prefix.
+        super().__init__(*args, **kwargs)
+
+        cmd = [f"{self.import_call_bin_prefix}neo4j-admin", "version"]
+        version = subprocess.check_output(cmd)
+        major = int(version.decode().split()[1].split(".")[0])
+        if major >= 5:
+            self.import_cmd = "database import full"
+            self.database_cmd = ""
+            self.wipe_cmd = "--overwrite-destination="
+        else:
+            self.import_cmd = "import"
+            self.database_cmd = "--database="
+            self.wipe_cmd = "--force="
+
     def _get_default_import_call_bin_prefix(self):
         """
         Method to provide the default string for the import call bin prefix.
@@ -1286,8 +1312,7 @@ class _Neo4jBatchWriter(_BatchWriter):
             str: a bash command for neo4j-admin import
         """
         import_call = (
-            f"{self.import_call_bin_prefix}neo4j-admin import "
-            f"--database={self.db_name} "
+            f"{self.import_call_bin_prefix}neo4j-admin {self.import_cmd} "
             f'--delimiter="{self.escaped_delim}" '
             f'--array-delimiter="{self.escaped_adelim}" '
         )
@@ -1298,7 +1323,7 @@ class _Neo4jBatchWriter(_BatchWriter):
             import_call += f"--quote='{self.quote}' "
 
         if self.wipe:
-            import_call += f"--force=true "
+            import_call += f"{self.wipe_cmd}true "
         if self.skip_bad_relationships:
             import_call += "--skip-bad-relationships=true "
         if self.skip_duplicate_nodes:
@@ -1312,6 +1337,8 @@ class _Neo4jBatchWriter(_BatchWriter):
         for header_path, parts_path in self.import_call_edges:
             import_call += f'--relationships="{header_path},{parts_path}" '
 
+        # Database needs to be at the end starting with Neo4j 5.0+.
+        import_call += f"{self.database_cmd}{self.db_name} "
         return import_call
 
 
