+++ test/test_write_neo4j.py
@@ -57,8 +57,10 @@ def test_create_import_call(bw):
     tmp_path = bw.outdir
 
     assert passed
-    assert 'bin/neo4j-admin import --database=neo4j --delimiter=";" ' in call
-    assert '--array-delimiter="|" --quote="\'" --force=true ' in call
+    assert 'neo4j-admin' in call
+    assert 'import' in call
+    assert '--delimiter=";"' in call
+    assert '--array-delimiter="|" --quote="\'"' in call
     assert (
         f'--nodes="{tmp_path}{os.sep}PostTranslationalInteraction-header.csv,{tmp_path}{os.sep}PostTranslationalInteraction-part.*" '
         in call
@@ -105,16 +107,15 @@ def test_neo4j_write_node_data_headers_import_call(bw, _get_nodes):
         protein_header
         == ":ID;name;score:double;taxon:long;genes:string[];id;preferred_id;:LABEL"
     )
-    assert micro_rna_header == ":ID;name;taxon:long;id;preferred_id;:LABEL"
-    assert "bin/neo4j-admin import" in call
-    assert "--database=neo4j" in call
-    assert '--delimiter=";"' in call
-    assert "--force=true" in call
-    assert '--nodes="' in call
-    assert "Protein-header.csv" in call
-    assert 'Protein-part.*"' in call
-    assert "MicroRNA-header.csv" in call
-    assert 'MicroRNA-part.*"' in call
+    assert m == ":ID;name;taxon:long;id;preferred_id;:LABEL"
+    assert "neo4j-admin" in c
+    assert "import" in c
+    assert '--delimiter=";"' in c
+    assert '--nodes="' in c
+    assert "Protein-header.csv" in c
+    assert 'Protein-part.*"' in c
+    assert "MicroRNA-header.csv" in c
+    assert 'MicroRNA-part.*"' in c
 
     # custom import call executable path
     bw.import_call_bin_prefix = "custom/path/"
@@ -125,7 +126,8 @@ def test_neo4j_write_node_data_headers_import_call(bw, _get_nodes):
     with open(import_call_path) as f:
         call = f.read()
 
-    assert "custom/path/neo4j-admin import" in call
+    assert "custom/path/neo4j-admin" in c
+    assert "import" in c
 
     # custom file prefix
     # TODO
@@ -816,10 +818,9 @@ def test_write_edge_data_headers_import_call(bw, _get_nodes, _get_edges):
     )
     assert is_mutated_in == ":START_ID;id;site;confidence:long;:END_ID;:TYPE"
 
-    assert "bin/neo4j-admin import" in call
-    assert "--database=neo4j" in call
+    assert "neo4j-admin" in call
+    assert "import" in call
     assert '--delimiter=";"' in call
-    assert "--force=true" in call
     assert '--nodes="' in call
     assert "PERTURBED_IN_DISEASE" in call
     assert "Is_Mutated_In" in call
