### 2023-09-18 39cee70
refactor(write): return the path of the import call file instead of a bool
This allows the user to manipulate the actual path, since no accessors exists otherwise.
For instance, this can be printed on stdout and piped in a shell automatically.
The returned string can still be used in a test, hence not changing the interface much.
