### 2023-09-22 9c39b16
issue #195: test mac os build
