### 2023-09-22 84bfe0d
issue #195: try to fix windows build
