### 2023-09-22 e1471cc
issue #195: windows build, use other neo4j docker image
