### 2023-09-22 1974717
issue #195: windows build, another try to use other neo4j docker image
