### 2023-09-25 862e552
issue #195: fix windows build by removing neo4j Docker setup on Windows
