### 2023-09-25 d220a7d
issue #195: fix windows build by removing postgres Docker setup on Windows
