### 2023-09-25 3574d4c
issue #195: fix windows build by correct pytest command
