### 2023-09-25 cb9594d
issue #259: run release only on tags; test all supported python versions
