+++ .github/workflows/ci_cd.yaml
@@ -1,22 +1,21 @@
 name: Tests and code quality
 
-on:
-  pull_request:
-  push:
-    branches:
-      - '*'
+on: [push, pull_request]
 
 jobs:
   build-linux:
     runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ["3.9", "3.10"]
 
     steps:
     - name: Checkout Repository
       uses: actions/checkout@v4
-    - name: Set up Python
+    - name: Set up Python ${{ matrix.python-version }}
       uses: actions/setup-python@v4
       with:
-        python-version: '3.9'
+        python-version: ${{ matrix.python-version }}
     - name: Install and configure Poetry
       uses: snok/install-poetry@v1
       with:
@@ -68,38 +67,38 @@ jobs:
     - name: Run Tests
       run: pytest --password=your_password_here
 
-  build-windows:
-    runs-on: windows-latest
-
-    defaults:
-      run:
-        shell: bash
-
-    steps:
-    - name: Checkout Repository
-      uses: actions/checkout@v4
-    - name: Set up Python
-      uses: actions/setup-python@v4
-      with:
-        python-version: '3.9'
-    - name: Install Poetry
-      uses: snok/install-poetry@v1
-      with:
-        virtualenvs-create: true
-        virtualenvs-in-project: true
-    - name: Load cached venv
-      id: cached-pip-wheels
-      uses: actions/cache@v3
-      with:
-        path: ~/.cache
-        key: venv-${{ runner.os }}-${{ steps.setup-python.outputs.python-version }}-${{ hashFiles('**/poetry.lock') }}
-    - name: Install Dependencies
-      run: poetry install --no-interaction --no-root
-    # - name: Start Neo4j Docker
-    # No official Neo4j Docker image available for Windows
-    # -> setup currently missing and tests which need this are skipped on Windows
-    # and the same for Postgres
-    # - name: Start Postgres Docker
-    #  run: docker run --restart always --publish=5432:5432 --env POSTGRES_PASSWORD=postgres -d postgres:latest
-    - name: Run Tests
-      run: poetry run pytest --password=your_password_here
+  #build-windows:
+  #  runs-on: windows-latest
+#
+  #  defaults:
+  #    run:
+  #      shell: bash
+#
+  #  steps:
+  #  - name: Checkout Repository
+  #    uses: actions/checkout@v4
+  #  - name: Set up Python
+  #    uses: actions/setup-python@v4
+  #    with:
+  #      python-version: '3.9'
+  #  - name: Install Poetry
+  #    uses: snok/install-poetry@v1
+  #    with:
+  #      virtualenvs-create: true
+  #      virtualenvs-in-project: true
+  #  - name: Load cached venv
+  #    id: cached-pip-wheels
+  #    uses: actions/cache@v3
+  #    with:
+  #      path: ~/.cache
+  #      key: venv-${{ runner.os }}-${{ steps.setup-python.outputs.python-version }}-${{ hashFiles('**/poetry.lock') }}
+  #  - name: Install Dependencies
+  #    run: poetry install --no-interaction --no-root
+  #  # - name: Start Neo4j Docker
+  #  # No official Neo4j Docker image available for Windows
+  #  # -> setup currently missing and tests which need this are skipped on Windows
+  #  # and the same for Postgres
+  #  # - name: Start Postgres Docker
+  #  #  run: docker run --restart always --publish=5432:5432 --env POSTGRES_PASSWORD=postgres -d postgres:latest
+  #  - name: Run Tests
+  #    run: poetry run pytest --password=your_password_here
