+++ .github/workflows/release.yaml
@@ -11,7 +11,7 @@ on:
 jobs:
     build_and_deploy_docs:
         runs-on: ubuntu-latest
-        if: github.ref == 'refs/heads/main'
+        if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags')
         permissions: write-all
         steps:
         -   name: Check out main
