### 2023-09-25 f1144b1
issue #259: use matrix config to test different os and python versions
