+++ .github/workflows/ci_cd.yaml
@@ -3,11 +3,13 @@ name: Tests and code quality
 on: [push, pull_request]
 
 jobs:
-  build-linux:
-    runs-on: ubuntu-latest
+  test:
     strategy:
       matrix:
+        fail-fast: true
+        os: ["ubuntu-latest", "macos-latest"]
         python-version: ["3.9", "3.10"]
+    runs-on: ${{ matrix.os }}
 
     steps:
     - name: Checkout Repository
@@ -16,6 +18,7 @@ jobs:
       uses: actions/setup-python@v4
       with:
         python-version: ${{ matrix.python-version }}
+        cache: 'poetry'
     - name: Install and configure Poetry
       uses: snok/install-poetry@v1
       with:
@@ -23,49 +26,55 @@ jobs:
         virtualenvs-create: false
     - name: Install Dependencies
       run: poetry install
-    - name: Start Neo4j Docker
-      run: docker run --restart always --publish=7474:7474 --publish=7687:7687 --env NEO4J_AUTH=neo4j/your_password_here --env NEO4J_PLUGINS='["apoc"]' --env=NEO4J_ACCEPT_LICENSE_AGREEMENT=yes -d neo4j:4.4-enterprise
-    - name: Start Postgres Docker
-      run: docker run --restart always --publish=5432:5432 --env POSTGRES_PASSWORD=postgres -d postgres:latest
-    - name: Run Tests
-      run: coverage run -m pytest --password=your_password_here
-    - name: Generate coverage badge
-      run: coverage-badge -f -o docs/coverage/coverage.svg
-    - name: Check code quality
-      uses: pre-commit/action@v3.0.0
-    - name: Commit changes
-      uses: s0/git-publish-subdir-action@develop
-      env:
-        REPO: self
-        BRANCH: coverage
-        FOLDER: docs/coverage
-        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-
-  build-macos:
-    runs-on: macos-latest
-
-    steps:
-    - name: Checkout Repository
-      uses: actions/checkout@v4
-    - name: Set up Python
-      uses: actions/setup-python@v4
-      with:
-        python-version: '3.9'
-    - name: Install and configure Poetry
-      uses: snok/install-poetry@v1
-      with:
-        version: 1.5.1
-        virtualenvs-create: false
-    - name: Install Dependencies
-      run: poetry install
-    - name: Install Docker
-      uses: douglascamata/setup-docker-macos-action@v1-alpha
     - name: Start Neo4j Docker
       run: docker run --restart always --publish=7474:7474 --publish=7687:7687 --env NEO4J_AUTH=neo4j/your_password_here --env NEO4J_PLUGINS='["apoc"]' --env=NEO4J_ACCEPT_LICENSE_AGREEMENT=yes -d neo4j:4.4-enterprise
     - name: Start Postgres Docker
       run: docker run --restart always --publish=5432:5432 --env POSTGRES_PASSWORD=postgres -d postgres:latest
     - name: Run Tests
       run: pytest --password=your_password_here
+    - name: Check code quality
+      uses: pre-commit/action@v3.0.0
+
+    - name: Generate coverage report
+      run: coverage run -m pytest --password=your_password_here
+      if: ${{ matrix.python-version == '3.9' && matrix.os == 'ubuntu-latest' }}
+    - name: Generate coverage badge
+      run: coverage-badge -f -o docs/coverage/coverage.svg
+      if: ${{ matrix.python-version == '3.9' && matrix.os == 'ubuntu-latest' }}
+    - name: Commit changes
[…]
