### 2023-09-25 7855bbd
issue #259: correct yaml syntax
