+++ .github/workflows/release.yaml
@@ -61,6 +61,7 @@ jobs:
 
     build_and_deploy_artifact:
         runs-on: ubuntu-latest
+        if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags')
         steps:
         -   name: Checkout Repository
             uses: actions/checkout@v4
