+++ .github/workflows/ci_cd.yaml
@@ -14,16 +14,18 @@ jobs:
     steps:
     - name: Checkout Repository
       uses: actions/checkout@v4
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v4
-      with:
-        python-version: ${{ matrix.python-version }}
-        cache: 'poetry'
     - name: Install and configure Poetry
       uses: snok/install-poetry@v1
       with:
         version: 1.5.1
         virtualenvs-create: false
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v4
+      with:
+        python-version: ${{ matrix.python-version }}
+        cache: 'poetry'
+    - name: Set Poetry Python version
+      run: poetry env use ${{ matrix.python-version }}
     - name: Install Dependencies
       run: poetry install
     - name: Start Neo4j Docker
