+++ .github/workflows/ci_cd.yaml
@@ -18,7 +18,7 @@ jobs:
       uses: snok/install-poetry@v1
       with:
         version: 1.5.1
-        virtualenvs-create: false
+        virtualenvs-create: true
     - name: Set up Python ${{ matrix.python-version }}
       uses: actions/setup-python@v4
       with:
