### 2023-09-25 7f78a18
issue #259: prefix commands with poetry run
