### 2023-09-25 9df66e1
issue #259: set concurrency to 1, due to neo4j/postgres databases
