### 2023-09-25 933d8a5
issue #259: fix macos docker setup; different push and pull request configs
