### 2023-09-25 a1bbe35
issue #259: fix syntax
