### 2023-09-25 48fcdff
issue #259: fix syntax
