### 2023-09-25 1513c10
issue #259: fix run conditions
