+++ .github/workflows/ci_cd.yaml
@@ -1,12 +1,6 @@
 name: Tests and code quality
 
-on:
-  pull_request:
-    branches:
-      - main
-  push:
-    branches:
-      - main
+on: [push, pull_request]
 
 jobs:
   push_job:
