+++ .github/workflows/ci_cd.yaml
@@ -6,7 +6,6 @@ jobs:
   push_job:
     if: ${{ github.event_name == 'push' }}
     runs-on: ubuntu-latest
-    # TODO: fix duplicate code
     steps:
     - name: Checkout Repository
       uses: actions/checkout@v4
@@ -34,9 +33,8 @@ jobs:
       uses: pre-commit/action@v3.0.0
 
   pull_request_job:
-    if: ${{ github.event_name == 'pull_request' }}
+    # if: ${{ github.event_name == 'pull_request' }}
     strategy:
-      # max-parallel: 1
       fail-fast: true
       matrix:
         os: ["ubuntu-latest", "macos-latest"]
@@ -148,3 +146,41 @@ jobs:
   #  #  run: docker run --restart always --publish=5432:5432 --env POSTGRES_PASSWORD=postgres -d postgres:latest
   #  - name: Run Tests
   #    run: poetry run pytest --password=your_password_here
+
+  build-windows:
+    runs-on: windows-latest
+
+    defaults:
+      run:
+        shell: bash
+
+    steps:
+    - name: Checkout Repository
+      uses: actions/checkout@v4
+    - name: Set up Python
+      uses: actions/setup-python@v4
+      with:
+        python-version: '3.9'
+    - name: Install Poetry
+      uses: snok/install-poetry@v1
+      with:
+        virtualenvs-create: true
+        virtualenvs-in-project: true
+    - name: Load cached venv
+      id: cached-pip-wheels
+      uses: actions/cache@v3
+      with:
+        path: ~/.cache
+        key: venv-${{ runner.os }}-${{ steps.setup-python.outputs.python-version }}-${{ hashFiles('**/poetry.lock') }}
+    - name: Install Dependencies
+      run: poetry install --no-interaction --no-root
+    # - name: Start Neo4j Docker
+    # No official Neo4j Docker image available for Windows
+    # -> setup currently missing and tests which need this are skipped on Windows
+    # and the same for Postgres
+    # - name: Start Postgres Docker
+    #  run: docker run --restart always --publish=5432:5432 --env POSTGRES_PASSWORD=postgres -d postgres:latest
+    - name: Run Tests
+      run: |
+          source $VENV
+          pytest --password=your_password_here
