### 2023-09-25 eb966c4
issue #259: another try to fix windows build
