### 2023-09-28 4b1b3be
issue #259: adapt tests to handle different path separators for windows and linux correctly
