+++ .github/workflows/release.yaml
@@ -60,8 +60,13 @@ jobs:
                 GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
 
     build_and_deploy_artifact:
-        runs-on: ubuntu-latest
         if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags')
+        runs-on: ubuntu-latest
+        environment:
+            name: pypi
+            url: https://test.pypi.org/project/biocypher/
+        permissions:
+            id-token: write
         steps:
         -   name: Checkout Repository
             uses: actions/checkout@v4
@@ -78,5 +83,5 @@ jobs:
             run: poetry install
         -   name: Build artifact
             run: poetry build
-        #- name: Deploy artifact to PyPi
-        #  run: poetry publish --username __token__ --password ${{ secrets.PYPI_TOKEN }}
+        -   name: Publish artifact to PyPI
+            uses: pypa/gh-action-pypi-publish@release/v1
