### 2023-09-28 8240d6d
issue #259: comment release part (not yet testable, because developed in forked repo)
