### 2023-09-28 5c5d89e
issue #259: try to set write permissions
