+++ .github/workflows/ci_cd.yaml
@@ -40,6 +40,8 @@ jobs:
         os: ["ubuntu-latest", "macos-latest"]
         python-version: ["3.9", "3.10", "3.11"]
     runs-on: ${{ matrix.os }}
+    permissions:
+      contents: write
 
     steps:
     - name: Checkout Repository
