### 2023-10-02 9b9c14a
issue #259: remove test coverage badge update from cicd workflow
