### 2023-10-02 b6a29f7
issue #195: include windws in the matrix
