+++ .github/workflows/ci_cd.yaml
@@ -37,13 +37,17 @@ jobs:
     strategy:
       fail-fast: true
       matrix:
-        os: ["ubuntu-latest", "macos-latest"]
+        os: ["ubuntu-latest", "macos-latest", "windows-latest"]
         python-version: ["3.9", "3.10", "3.11"]
     runs-on: ${{ matrix.os }}
-    permissions:
-      contents: write
+    defaults:
+      run:
+        shell: bash
 
     steps:
+    #----------------------------------------------
+    # check-out repo and set-up python and poetry
+    #----------------------------------------------
     - name: Checkout Repository
       uses: actions/checkout@v4
     - name: Install and configure Poetry
@@ -51,6 +55,7 @@ jobs:
       with:
         version: 1.5.1
         virtualenvs-create: true
+        virtualenvs-in-project: true
     - name: Set up Python ${{ matrix.python-version }}
       uses: actions/setup-python@v4
       with:
@@ -58,8 +63,24 @@ jobs:
         cache: 'poetry'
     - name: Set Poetry Python version
       run: poetry env use ${{ matrix.python-version }}
-    - name: Install Dependencies
-      run: poetry install
+    - name: Load cached venv
+      if: ${{ matrix.os == 'windows-latest' }}
+      id: cached-pip-wheels
+      uses: actions/cache@v3
+      with:
+        path: ~/.cache
+        key: venv-${{ runner.os }}-${{ steps.setup-python.outputs.python-version }}-${{ hashFiles('**/poetry.lock') }}
+    #----------------------------------------------
+    #             install dependencies
+    #----------------------------------------------
+    - name: Install dependencies
+      run: poetry install --no-interaction --no-root
+    - name: Install library
+      run: poetry install --no-interaction
+    #----------------------------------------------
+    #    setup docker containers for testing
+    #----------------------------------------------
+    # currently only available for ubuntu and macos
     - name: Install Docker
       uses: douglascamata/setup-docker-macos-action@v1-alpha
       if: ${{ matrix.os == 'macos-latest' }}
@@ -67,8 +88,14 @@ jobs:
       run: docker run --restart always --publish=7474:7474 --publish=7687:7687 --env NEO4J_AUTH=neo4j/your_password_here --env NEO4J_PLUGINS='["apoc"]' --env=NEO4J_ACCEPT_LICENSE_AGREEMENT=yes -d neo4j:4.4-enterprise
     - name: Start Postgres Docker
       run: docker run --restart always --publish=5432:5432 --env POSTGRES_PASSWORD=postgres -d postgres:latest
+    #----------------------------------------------
+    #      run tests and code quality checks
+    #----------------------------------------------
     - name: Run Tests
-      run: poetry run pytest --password=your_password_here
+      run: |
+        source $VENV
+        pytest --password=your_password_here
+      # run: poetry run pytest --password=your_password_here
     - name: Check code quality
       uses: pre-commit/action@v3.0.0
 
@@ -87,42 +114,42 @@ jobs:
     #    FOLDER: docs/coverage
     #    GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
 
-  build-windows:
-    if: ${{ github.event_name == 'pull_request' }}
-    runs-on: windows-latest
-    defaults:
-      run:
-        shell: bash
+  #build-windows:
+  #  if: ${{ github.event_name == 'pull_request' }}
+  #  runs-on: windows-latest
+  #  defaults:
+  #    run:
+  #      shell: bash
 
-    steps:
-    - name: Checkout Repository
-      uses: actions/checkout@v4
-    - name: Set up Python
-      uses: actions/setup-python@v4
-      with:
-        python-version: '3.9'
[…]
