### 2023-10-02 699db47
issue #195: adapt setup section for usage with windows
