+++ .github/workflows/ci_cd.yaml
@@ -50,21 +50,19 @@ jobs:
     #----------------------------------------------
     - name: Checkout Repository
       uses: actions/checkout@v4
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v4
+      with:
+        python-version: ${{ matrix.python-version }}
     - name: Install and configure Poetry
       uses: snok/install-poetry@v1
       with:
         version: 1.5.1
         virtualenvs-create: true
         virtualenvs-in-project: true
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v4
-      with:
-        python-version: ${{ matrix.python-version }}
-        cache: 'poetry'
-    - name: Set Poetry Python version
-      run: poetry env use ${{ matrix.python-version }}
+    #- name: Set Poetry Python version
+    #  run: poetry env use ${{ matrix.python-version }}
     - name: Load cached venv
-      if: ${{ matrix.os == 'windows-latest' }}
       id: cached-pip-wheels
       uses: actions/cache@v3
       with:
