### 2023-10-02 fd27978
issue #195: do not run docker section on windows
