### 2023-10-02 2f30aac
issue #195: refactoring and clean up; issue #257 set postgres version in cicd pipeline and not depend on latest tag
