+++ .github/workflows/ci_cd.yaml
@@ -6,27 +6,40 @@ jobs:
   push_job:
     if: ${{ github.event_name == 'push' }}
     runs-on: ubuntu-latest
+    env:
+      PYTHON_VERSION: 3.9
+
     steps:
+    #----------------------------------------------
+    # check-out repo and set-up python and poetry
+    #----------------------------------------------
     - name: Checkout Repository
       uses: actions/checkout@v4
+    - name: Set up Python $PYTHON_VERSION
+      uses: actions/setup-python@v4
+      with:
+        python-version: $PYTHON_VERSION
     - name: Install and configure Poetry
       uses: snok/install-poetry@v1
       with:
         version: 1.5.1
         virtualenvs-create: true
-    - name: Set up Python 3.9
-      uses: actions/setup-python@v4
-      with:
-        python-version: 3.9
-        cache: 'poetry'
-    - name: Set Poetry Python version
-      run: poetry env use 3.9
+        virtualenvs-in-project: true
+    #----------------------------------------------
+    #             install dependencies
+    #----------------------------------------------
     - name: Install Dependencies
       run: poetry install
+    #----------------------------------------------
+    #    setup docker containers for testing
+    #----------------------------------------------
     - name: Start Neo4j Docker
       run: docker run --restart always --publish=7474:7474 --publish=7687:7687 --env NEO4J_AUTH=neo4j/your_password_here --env NEO4J_PLUGINS='["apoc"]' --env=NEO4J_ACCEPT_LICENSE_AGREEMENT=yes -d neo4j:4.4-enterprise
     - name: Start Postgres Docker
-      run: docker run --restart always --publish=5432:5432 --env POSTGRES_PASSWORD=postgres -d postgres:latest
+      run: docker run --restart always --publish=5432:5432 --env POSTGRES_PASSWORD=postgres -d postgres:11.21-bullseye
+    #----------------------------------------------
+    #      run tests and code quality checks
+    #----------------------------------------------
     - name: Run Tests
       run: poetry run pytest --password=your_password_here
     - name: Check code quality
@@ -86,7 +99,7 @@ jobs:
       run: docker run --restart always --publish=7474:7474 --publish=7687:7687 --env NEO4J_AUTH=neo4j/your_password_here --env NEO4J_PLUGINS='["apoc"]' --env=NEO4J_ACCEPT_LICENSE_AGREEMENT=yes -d neo4j:4.4-enterprise
       if: ${{ matrix.os != 'windows-latest' }}
     - name: Start Postgres Docker
-      run: docker run --restart always --publish=5432:5432 --env POSTGRES_PASSWORD=postgres -d postgres:latest
+      run: docker run --restart always --publish=5432:5432 --env POSTGRES_PASSWORD=postgres -d postgres:11.21-bullseye
       if: ${{ matrix.os != 'windows-latest' }}
     #----------------------------------------------
     #      run tests and code quality checks
@@ -95,10 +108,8 @@ jobs:
       run: |
         source $VENV
         pytest --password=your_password_here
-      # run: poetry run pytest --password=your_password_here
     - name: Check code quality
       uses: pre-commit/action@v3.0.0
-
     - name: Generate coverage report
       run: poetry run coverage run -m pytest --password=your_password_here
       if: ${{ matrix.python-version == '3.9' && matrix.os == 'ubuntu-latest' }}
@@ -113,43 +124,3 @@ jobs:
     #    BRANCH: coverage
     #    FOLDER: docs/coverage
     #    GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-
-  #build-windows:
-  #  if: ${{ github.event_name == 'pull_request' }}
[…]
