### 2023-10-02 c9f4b70
issue #195: fix python version env variable syntax
