### 2023-10-02 f7d4e81
issue #259: separate docs and release workflows; build docs on every commit
