+++ .github/workflows/docs.yaml
@@ -1,6 +1,6 @@
 name: Docs
 
-on: [push]
+on: [push, pull_request]
 
 jobs:
   build_docs:
