### 2023-10-10 a3d0a7f
issue #259: make autopublish testable:  configure publish workflow to publish to staging (test PyPi); update test coverage to be testable on pyopensci branch
