### 2023-10-11 79e280b
issue #254: replace single-letter variables in tests
