+++ test/test_write_arango.py
@@ -14,75 +14,81 @@ def test_arango_write_data_headers_import_call(
 
     edges = _get_edges
 
-    p1 = bw_arango.write_nodes(nodes[:4])
-    p2 = bw_arango.write_nodes(nodes[4:])
-    p3 = bw_arango.write_edges(edges[:4])
-    p4 = bw_arango.write_edges(edges[4:])
+    protein_1 = bw_arango.write_nodes(nodes[:4])
+    protein_2 = bw_arango.write_nodes(nodes[4:])
+    protein_3 = bw_arango.write_edges(edges[:4])
+    protein_4 = bw_arango.write_edges(edges[4:])
 
-    assert all([p1, p2, p3, p4])
+    assert all([protein_1, protein_2, protein_3, protein_4])
 
     bw_arango.write_import_call()
 
     tmp_path = bw_arango.outdir
 
-    ph_csv = os.path.join(tmp_path, "Protein-header.csv")
-    pp_1_csv = os.path.join(tmp_path, "Protein-part000.csv")
-    pp_2_csv = os.path.join(tmp_path, "Protein-part001.csv")
-    mh_csv = os.path.join(tmp_path, "MicroRNA-header.csv")
-    mp_1_csv = os.path.join(tmp_path, "MicroRNA-part000.csv")
-    mp_2_csv = os.path.join(tmp_path, "MicroRNA-part001.csv")
-    dh_csv = os.path.join(tmp_path, "PERTURBED_IN_DISEASE-header.csv")
-    dp_1_csv = os.path.join(tmp_path, "PERTURBED_IN_DISEASE-part000.csv")
-    dp_2_csv = os.path.join(tmp_path, "PERTURBED_IN_DISEASE-part001.csv")
-    muh_csv = os.path.join(tmp_path, "Is_Mutated_In-header.csv")
-    mup_1_csv = os.path.join(tmp_path, "Is_Mutated_In-part000.csv")
-    mup_2_csv = os.path.join(tmp_path, "Is_Mutated_In-part001.csv")
+    protein_header_csv = os.path.join(tmp_path, "Protein-header.csv")
+    protein_data_1_csv = os.path.join(tmp_path, "Protein-part000.csv")
+    protein_data_2_csv = os.path.join(tmp_path, "Protein-part001.csv")
+    micro_rna_header_csv = os.path.join(tmp_path, "MicroRNA-header.csv")
+    micro_rna_data_1_csv = os.path.join(tmp_path, "MicroRNA-part000.csv")
+    micro_rna_data_2_csv = os.path.join(tmp_path, "MicroRNA-part001.csv")
+    disease_header_csv = os.path.join(
+        tmp_path, "PERTURBED_IN_DISEASE-header.csv"
+    )
+    disease_data_1_csv = os.path.join(
+        tmp_path, "PERTURBED_IN_DISEASE-part000.csv"
+    )
+    disease_data_2_csv = os.path.join(
+        tmp_path, "PERTURBED_IN_DISEASE-part001.csv"
+    )
+    mutation_header_csv = os.path.join(tmp_path, "Is_Mutated_In-header.csv")
+    mutation_data_1_csv = os.path.join(tmp_path, "Is_Mutated_In-part000.csv")
+    mutation_data_2_csv = os.path.join(tmp_path, "Is_Mutated_In-part001.csv")
     call_csv = os.path.join(tmp_path, "arangodb-import-call.sh")
 
-    with open(ph_csv) as f:
-        ph = f.read()
-    with open(pp_1_csv) as f:
-        pp_1 = f.readlines()
-    with open(pp_2_csv) as f:
-        pp_2 = f.readlines()
-    with open(mh_csv) as f:
[…]
