+++ test/test_write_neo4j.py
@@ -19,30 +19,30 @@ def test_neo4j_writer_and_output_dir(bw):
 
 def test_create_import_call(bw):
     mixed = []
-    le = 4
-    for i in range(le):
-        n = BioCypherNode(
+    number_of_items = 4
+    for i in range(number_of_items):
+        node = BioCypherNode(
             f"i{i+1}",
             "post translational interaction",
         )
-        e1 = BioCypherEdge(
+        edge_1 = BioCypherEdge(
             source_id=f"i{i+1}",
             target_id=f"p{i+1}",
             relationship_label="IS_SOURCE_OF",
         )
-        e2 = BioCypherEdge(
+        edge_2 = BioCypherEdge(
             source_id=f"i{i}",
             target_id=f"p{i+2}",
             relationship_label="IS_TARGET_OF",
         )
-        mixed.append(BioCypherRelAsNode(n, e1, e2))
+        mixed.append(BioCypherRelAsNode(node, edge_1, edge_2))
 
-        e3 = BioCypherEdge(
+        edge_3 = BioCypherEdge(
             source_id=f"p{i+1}",
             target_id=f"p{i+1}",
             relationship_label="PERTURBED_IN_DISEASE",
         )
-        mixed.append(e3)
+        mixed.append(edge_3)
 
     def gen(lis):
         yield from lis
@@ -87,42 +87,42 @@ def test_neo4j_write_node_data_headers_import_call(bw, _get_nodes):
 
     tmp_path = bw.outdir
 
-    p_csv = os.path.join(tmp_path, "Protein-header.csv")
-    m_csv = os.path.join(tmp_path, "MicroRNA-header.csv")
-    call = os.path.join(tmp_path, "neo4j-admin-import-call.sh")
+    protein_header_csv = os.path.join(tmp_path, "Protein-header.csv")
+    micro_rna_header_csv = os.path.join(tmp_path, "MicroRNA-header.csv")
+    import_call_path = os.path.join(tmp_path, "neo4j-admin-import-call.sh")
 
-    with open(p_csv) as f:
-        p = f.read()
-    with open(m_csv) as f:
-        m = f.read()
-    with open(call) as f:
-        c = f.read()
+    with open(protein_header_csv) as f:
+        protein_header = f.read()
+    with open(micro_rna_header_csv) as f:
+        micro_rna_header = f.read()
+    with open(import_call_path) as f:
+        call = f.read()
 
     assert (
-        p
+        protein_header
         == ":ID;name;score:double;taxon:long;genes:string[];id;preferred_id;:LABEL"
     )
-    assert m == ":ID;name;taxon:long;id;preferred_id;:LABEL"
-    assert "bin/neo4j-admin import" in c
-    assert "--database=neo4j" in c
-    assert '--delimiter=";"' in c
-    assert "--force=true" in c
-    assert '--nodes="' in c
-    assert "Protein-header.csv" in c
-    assert 'Protein-part.*"' in c
-    assert "MicroRNA-header.csv" in c
-    assert 'MicroRNA-part.*"' in c
+    assert micro_rna_header == ":ID;name;taxon:long;id;preferred_id;:LABEL"
+    assert "bin/neo4j-admin import" in call
+    assert "--database=neo4j" in call
+    assert '--delimiter=";"' in call
+    assert "--force=true" in call
+    assert '--nodes="' in call
+    assert "Protein-header.csv" in call
+    assert 'Protein-part.*"' in call
+    assert "MicroRNA-header.csv" in call
[…]
