+++ test/test_integration.py
@@ -15,20 +15,23 @@ def test_write_node_data_from_gen(core, _get_nodes):
 
     path = core._output_directory
 
-    p_csv = os.path.join(path, "Protein-part000.csv")
-    m_csv = os.path.join(path, "MicroRNA-part000.csv")
+    protein_csv = os.path.join(path, "Protein-part000.csv")
+    micro_rna_csv = os.path.join(path, "MicroRNA-part000.csv")
 
-    with open(p_csv) as f:
-        pr = f.read()
+    with open(protein_csv) as f:
+        protein_data = f.read()
 
-    with open(m_csv) as f:
-        mi = f.read()
+    with open(micro_rna_csv) as f:
+        micro_rna_data = f.read()
 
     assert passed
-    assert "p1;'StringProperty1';4.0;9606;'gene1|gene2';'p1';'uniprot'" in pr
-    assert "BiologicalEntity" in pr
-    assert "m1;'StringProperty1';9606;'m1';'mirbase'" in mi
-    assert "ChemicalEntity" in mi
+    assert (
+        "p1;'StringProperty1';4.0;9606;'gene1|gene2';'p1';'uniprot'"
+        in protein_data
+    )
+    assert "BiologicalEntity" in protein_data
+    assert "m1;'StringProperty1';9606;'m1';'mirbase'" in micro_rna_data
+    assert "ChemicalEntity" in micro_rna_data
 
 
 def test_show_ontology_structure_kwargs(core):
