### 2023-10-11 ef74095
issue #254: replace single-letter variables in source code and verify by running flake8 and check for E741
