+++ test/test_write_neo4j.py
@@ -74,7 +74,7 @@ def test_create_import_call(bw):
     )
 
 
-@pytest.mark.parametrize("l", [4], scope="module")
+@pytest.mark.parametrize("lenght", [4], scope="module")
 def test_neo4j_write_node_data_headers_import_call(bw, _get_nodes):
     # four proteins, four miRNAs
     nodes = _get_nodes
@@ -197,7 +197,7 @@ def test_property_types(bw):
     assert "BiologicalEntity" in data
 
 
-@pytest.mark.parametrize("l", [4], scope="module")
+@pytest.mark.parametrize("lenght", [4], scope="module")
 def test_write_node_data_from_list(bw, _get_nodes):
     nodes = _get_nodes
 
@@ -223,7 +223,7 @@ def test_write_node_data_from_list(bw, _get_nodes):
     assert "ChemicalEntity" in micro_rna
 
 
-@pytest.mark.parametrize("l", [4], scope="module")
+@pytest.mark.parametrize("lenght", [4], scope="module")
 def test_write_node_data_from_gen(bw, _get_nodes):
     nodes = _get_nodes
 
@@ -297,7 +297,7 @@ def test_write_node_data_from_gen_no_props(bw):
     assert "ChemicalEntity" in micro_rna
 
 
-@pytest.mark.parametrize("l", [int(1e4 + 4)], scope="module")
+@pytest.mark.parametrize("lenght", [int(1e4 + 4)], scope="module")
 def test_write_node_data_from_large_gen(bw, _get_nodes):
     nodes = _get_nodes
 
@@ -330,7 +330,7 @@ def test_write_node_data_from_large_gen(bw, _get_nodes):
     )
 
 
-@pytest.mark.parametrize("l", [1], scope="module")
+@pytest.mark.parametrize("lenght", [1], scope="module")
 def test_too_many_properties(bw, _get_nodes):
     nodes = _get_nodes
 
@@ -357,7 +357,7 @@ def test_too_many_properties(bw, _get_nodes):
     assert not passed
 
 
-@pytest.mark.parametrize("l", [1], scope="module")
+@pytest.mark.parametrize("lenght", [1], scope="module")
 def test_not_enough_properties(bw, _get_nodes):
     nodes = _get_nodes
 
@@ -439,7 +439,7 @@ def test_write_none_type_property_and_order_invariance(bw):
     assert "BiologicalEntity" in protein
 
 
-@pytest.mark.parametrize("l", [int(1e4)], scope="module")
+@pytest.mark.parametrize("lenght", [int(1e4)], scope="module")
 def test_accidental_exact_batch_size(bw, _get_nodes):
     nodes = _get_nodes
 
@@ -481,7 +481,7 @@ def test_accidental_exact_batch_size(bw, _get_nodes):
     )
 
 
-@pytest.mark.parametrize("l", [4], scope="module")
+@pytest.mark.parametrize("lenght", [4], scope="module")
 def test_write_edge_data_from_gen(bw, _get_edges):
     edges = _get_edges
 
@@ -529,7 +529,7 @@ def test_write_edge_data_from_gen(bw, _get_edges):
     assert "\n" in is_mutated_in
 
 
-@pytest.mark.parametrize("l", [int(1e4 + 4)], scope="module")
+@pytest.mark.parametrize("lenght", [int(1e4 + 4)], scope="module")
 def test_write_edge_data_from_large_gen(bw, _get_edges):
     edges = _get_edges
 
@@ -567,7 +567,7 @@ def test_write_edge_data_from_large_gen(bw, _get_edges):
     )
 
 
-@pytest.mark.parametrize("l", [4], scope="module")
+@pytest.mark.parametrize("lenght", [4], scope="module")
[…]
