### 2023-10-11 62ee0f6
issue #256: replace globals by fixtures
