+++ .github/workflows/ci_cd.yaml
@@ -33,9 +33,19 @@ jobs:
         virtualenvs-create: true
         virtualenvs-in-project: true
     #----------------------------------------------
-    #             install dependencies
+    #       load cached venv if cache exists
+    #----------------------------------------------
+    - name: Load cached venv
+      id: cached-poetry-dependencies
+      uses: actions/cache@v3
+      with:
+        path: .venv
+        key: venv-${{ runner.os }}-${{ steps.setup-python.outputs.python-version }}-${{ hashFiles('**/poetry.lock') }}
+    #----------------------------------------------
+    # install dependencies if cache does not exist
     #----------------------------------------------
     - name: Install Dependencies
+      if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
       run: poetry install
     #----------------------------------------------
     #    setup docker containers for testing
@@ -87,18 +97,20 @@ jobs:
         version: 1.5.1
         virtualenvs-create: true
         virtualenvs-in-project: true
-    #- name: Set Poetry Python version
-    #  run: poetry env use ${{ matrix.python-version }}
+    #----------------------------------------------
+    #       load cached venv if cache exists
+    #----------------------------------------------
     - name: Load cached venv
-      id: cached-pip-wheels
+      id: cached-poetry-dependencies
       uses: actions/cache@v3
       with:
-        path: ~/.cache
+        path: .venv
         key: venv-${{ runner.os }}-${{ steps.setup-python.outputs.python-version }}-${{ hashFiles('**/poetry.lock') }}
     #----------------------------------------------
-    #             install dependencies
+    # install dependencies if cache does not exist
     #----------------------------------------------
     - name: Install dependencies
+      if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
       run: poetry install --no-interaction --no-root
     - name: Install library
       run: poetry install --no-interaction
