+++ .github/workflows/reusable_steps/setup.yaml
@@ -1,36 +1,39 @@
-name: Setup
-
-run-name: ${{ github.actor }} is testing out GitHub Actions
-
-on:
-  workflow_call:
-
-jobs:
-  runs-on: ubuntu-latest
+name: 'Setup'
+description: 'Setup Python and Poetry'
+#inputs:
+#  who-to-greet:  # id of input
+#    description: 'Who to greet'
+#    required: true
+#    default: 'World'
+#outputs:
+#  random-number:
+#    description: "Random number"
+#    value: ${{ steps.random-number-generator.outputs.random-number }}
+runs:
+  using: "composite"
   env:
     PYTHON_VERSION: 3.9
     POETRY_VERSION: 1.5.1
-
   steps:
-  #----------------------------------------------
-  # check-out repo and set-up python and poetry
-  #----------------------------------------------
-  - name: Checkout Repository
-    uses: actions/checkout@v4
-  - name: Set up Python ${{env.PYTHON_VERSION}}
-    uses: actions/setup-python@v4
-    with:
-      python-version: ${{env.PYTHON_VERSION}}
-  - name: Load cached Poetry installation
-    id: cached-poetry
-    uses: actions/cache@v3
-    with:
-      path: ~/.local
-      key: poetry-cache-${{ runner.os }}-${{ env.PYTHON_VERSION }}-${{ env.POETRY_VERSION }}
-  - name: Install and configure Poetry
-    if: steps.cached-poetry.outputs.cache-hit != 'true'
-    uses: snok/install-poetry@v1
-    with:
-      version: 1.5.1
-      virtualenvs-create: true
-      virtualenvs-in-project: true
+    #----------------------------------------------
+    # check-out repo and set-up python and poetry
+    #----------------------------------------------
+    - name: Checkout Repository
+      uses: actions/checkout@v4
+    - name: Set up Python ${{env.PYTHON_VERSION}}
+      uses: actions/setup-python@v4
+      with:
+        python-version: ${{env.PYTHON_VERSION}}
+    - name: Load cached Poetry installation
+      id: cached-poetry
+      uses: actions/cache@v3
+      with:
+        path: ~/.local
+        key: poetry-cache-${{ runner.os }}-${{ env.PYTHON_VERSION }}-${{ env.POETRY_VERSION }}
+    - name: Install and configure Poetry
+      if: steps.cached-poetry.outputs.cache-hit != 'true'
+      uses: snok/install-poetry@v1
+      with:
+        version: 1.5.1
+        virtualenvs-create: true
+        virtualenvs-in-project: true
