### 2023-10-19 0f1d70f
issue #259: first checkout repo, then call setup.yaml
