+++ .github/workflows/ci_cd.yaml
@@ -8,6 +8,8 @@ jobs:
     runs-on: ubuntu-latest
 
     steps:
+    - name: Checkout Repository
+      uses: actions/checkout@v4
     - name: Setup
       uses: .github/workflows/reusable_steps/setup.yaml
     #----------------------------------------------
