+++ .github/workflows/ci_cd.yaml
@@ -11,7 +11,7 @@ jobs:
     - name: Checkout Repository
       uses: actions/checkout@v4
     - name: Setup
-      uses: .github/workflows/reusable_steps/setup.yaml
+      uses: ./.github/workflows/reusable_steps/setup.yaml
     #----------------------------------------------
     #       load cached venv if cache exists
     #----------------------------------------------
