### 2023-10-19 28c69d0
issue #259: fix folder layout to enable composite actions
