+++ .github/workflows/ci_cd.yaml
@@ -11,7 +11,7 @@ jobs:
     - name: Checkout Repository
       uses: actions/checkout@v4
     - name: Setup
-      uses: ./.github/actions/setup_composite/action.yaml
+      uses: nilskre/biocypher/.github/actions/setup_composite/action.yaml@PyOpenSci_review_part_2 # TODO: adapt
     #----------------------------------------------
     #       load cached venv if cache exists
     #----------------------------------------------
