### 2023-10-19 554d9e9
issue #259: set python and poetry version in composite by input
