+++ .github/workflows/ci_cd.yaml
@@ -8,11 +8,12 @@ jobs:
     runs-on: ubuntu-latest
 
     steps:
+    #----------------------------------------------
+    #  checkout repo and set-up python and poetry
+    #----------------------------------------------
     - name: Checkout Repository
       uses: actions/checkout@v4
-    #- name: Setup
-    #  uses: nilskre/biocypher/.github/actions/setup_composite/action.yaml@PyOpenSci_review_part_2 # TODO: adapt
-    - id: Setup
+    - name: Setup Python and Poetry
       uses: ./.github/actions/setup
     #----------------------------------------------
     #       load cached venv if cache exists
@@ -60,27 +61,15 @@ jobs:
 
     steps:
     #----------------------------------------------
-    # check-out repo and set-up python and poetry
+    #  checkout repo and set-up python and poetry
     #----------------------------------------------
     - name: Checkout Repository
       uses: actions/checkout@v4
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v4
+    - name: Setup Python and Poetry
+      uses: ./.github/actions/setup
       with:
-        python-version: ${{ matrix.python-version }}
-    - name: Load cached Poetry installation
-      id: cached-poetry
-      uses: actions/cache@v3
-      with:
-        path: ~/.local
-        key: poetry-cache-${{ runner.os }}-${{ matrix.python-version }}-${{ env.POETRY_VERSION }}
-    - name: Install and configure Poetry
-      if: steps.cached-poetry.outputs.cache-hit != 'true'
-      uses: snok/install-poetry@v1
-      with:
-        version: ${{env.POETRY_VERSION}}
-        virtualenvs-create: true
-        virtualenvs-in-project: true
+        PYTHON_VERSION: ${{ matrix.python-version }}
+        POETRY_VERSION: ${{ env.POETRY_VERSION }}
     #----------------------------------------------
     #       load cached venv if cache exists
     #----------------------------------------------
