+++ .github/actions/setup/action.yaml
@@ -3,23 +3,14 @@ description: 'Setup Python and Poetry'
 inputs:
   PYTHON_VERSION:
     description: 'Python version'
-    #required: true
     default: 3.9
   POETRY_VERSION:
     description: 'Poetry version'
-    #required: true
     default: 1.5.1
-#outputs:
-#  random-number:
-#    description: "Random number"
-#    value: ${{ steps.random-number-generator.outputs.random-number }}
 
 runs:
   using: "composite"
   steps:
-    #----------------------------------------------
-    # check-out repo and set-up python and poetry
-    #----------------------------------------------
     - name: Set up Python ${{inputs.PYTHON_VERSION}}
       uses: actions/setup-python@v4
       with:
@@ -37,3 +28,5 @@ runs:
         version: 1.5.1
         virtualenvs-create: true
         virtualenvs-in-project: true
+    - name: Check Poetry installation
+      run: poetry --version
