### 2023-10-19 6837070
issue #259: specify shell
