### 2023-10-19 37a59e8
added manual format specification, still need to allow not having a schema config
