### 2023-10-19 8b5f3d9
conditional ontology extension
