+++ .github/workflows/publish.yaml
@@ -3,6 +3,7 @@ name: Publish
 on: [push]
 
 jobs:
+  # TODO add 'main' branch filter
   build_and_deploy_artifact:
     if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags')
     runs-on: ubuntu-latest
@@ -43,6 +44,8 @@ jobs:
     #----------------------------------------------
     - name: Publish artifact to PyPI
       uses: pypa/gh-action-pypi-publish@release/v1
+      with:
+        repository_url: https://test.pypi.org/legacy/
     #----------------------------------------------
     #           create Github release
     #----------------------------------------------
