### 2023-10-19 fb53d1a
issue #259: add permission for github action to upload release to Github; fix testpypi repository-url
