### 2023-10-19 fda88c7
issue #259: remove duplicated code in cicd config by using composite actions
