+++ .github/actions/test/action.yaml
@@ -1,0 +1,28 @@
+name: 'Test and code quality'
+description: 'Run tests and code quality checks'
+
+runs:
+  using: "composite"
+  steps:
+    #----------------------------------------------
+    #    setup docker containers for testing
+    #----------------------------------------------
+    # currently only available for ubuntu and macos
+    - name: Install Docker
+      uses: douglascamata/setup-docker-macos-action@v1-alpha
+      if: ${{ runner.os == 'macos-latest' }}
+    - name: Start Neo4j Docker
+      run: docker run --restart always --publish=7474:7474 --publish=7687:7687 --env NEO4J_AUTH=neo4j/your_password_here --env NEO4J_PLUGINS='["apoc"]' --env=NEO4J_ACCEPT_LICENSE_AGREEMENT=yes -d neo4j:4.4-enterprise
+      if: ${{ runner.os != 'windows-latest' }}
+    - name: Start Postgres Docker
+      run: docker run --restart always --publish=5432:5432 --env POSTGRES_PASSWORD=postgres -d postgres:11.21-bullseye
+      if: ${{ runner.os != 'windows-latest' }}
+    #----------------------------------------------
+    #      run tests and code quality checks
+    #----------------------------------------------
+    - name: Run Tests
+      run: |
+        source $VENV
+        pytest --password=your_password_here
+    - name: Check code quality
+      uses: pre-commit/action@v3.0.0
