+++ .github/actions/build_docs/action.yaml
@@ -1,0 +1,10 @@
+name: 'Build docs'
+description: 'Build docs and run doc tests'
+
+runs:
+  using: "composite"
+  steps:
+    - name: Test code snippets in documentation
+      run: poetry run make doctest --directory docs/
+    - name: Build documentation
+      run: poetry run make html --directory docs/
