+++ .github/actions/test_coverage/action.yaml
@@ -1,0 +1,17 @@
+name: 'Test coverage'
+description: 'Check test coverage'
+
+runs:
+  using: "composite"
+  steps:
+    - name: Generate coverage report
+      run: poetry run coverage run -m pytest --password=your_password_here
+    - name: Generate coverage badge
+      run: poetry run coverage-badge -f -o docs/coverage/coverage.svg
+    - name: Commit changes
+      uses: s0/git-publish-subdir-action@develop
+      env:
+        REPO: self
+        BRANCH: coverage
+        FOLDER: docs/coverage
+        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
