### 2023-10-19 32be5f0
issue #259: specify shell and add main branch rule for publish workflow
