+++ .github/workflows/publish.yaml
@@ -1,9 +1,16 @@
 name: Publish
 
-on: [push]
+#on: [push]
+# TODO: switch to this rules after successful test
+on:
+  push:
+    branches:
+      - main
+    tags:
+    # TODO: ensure semver structure of tags
+      - '*'
 
 jobs:
-  # TODO add 'main' branch filter
   build_and_deploy_artifact:
     if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags')
     runs-on: ubuntu-latest
