+++ .github/actions/build_docs/action.yaml
@@ -6,5 +6,7 @@ runs:
   steps:
     - name: Test code snippets in documentation
       run: poetry run make doctest --directory docs/
+      shell: bash
     - name: Build documentation
       run: poetry run make html --directory docs/
+      shell: bash
