+++ .github/actions/test/action.yaml
@@ -13,9 +13,11 @@ runs:
       if: ${{ runner.os == 'macos-latest' }}
     - name: Start Neo4j Docker
       run: docker run --restart always --publish=7474:7474 --publish=7687:7687 --env NEO4J_AUTH=neo4j/your_password_here --env NEO4J_PLUGINS='["apoc"]' --env=NEO4J_ACCEPT_LICENSE_AGREEMENT=yes -d neo4j:4.4-enterprise
+      shell: bash
       if: ${{ runner.os != 'windows-latest' }}
     - name: Start Postgres Docker
       run: docker run --restart always --publish=5432:5432 --env POSTGRES_PASSWORD=postgres -d postgres:11.21-bullseye
+      shell: bash
       if: ${{ runner.os != 'windows-latest' }}
     #----------------------------------------------
     #      run tests and code quality checks
@@ -24,5 +26,6 @@ runs:
       run: |
         source $VENV
         pytest --password=your_password_here
+      shell: bash
     - name: Check code quality
       uses: pre-commit/action@v3.0.0
