### 2023-10-19 276d969
issue #259: check if tag corresponds to semver; correct runner.os checks
