+++ .github/actions/test/action.yaml
@@ -10,15 +10,15 @@ runs:
     # currently only available for ubuntu and macos
     - name: Install Docker
       uses: douglascamata/setup-docker-macos-action@v1-alpha
-      if: ${{ runner.os == 'macos-latest' }}
+      if: ${{ runner.os == 'macOS' }}
     - name: Start Neo4j Docker
       run: docker run --restart always --publish=7474:7474 --publish=7687:7687 --env NEO4J_AUTH=neo4j/your_password_here --env NEO4J_PLUGINS='["apoc"]' --env=NEO4J_ACCEPT_LICENSE_AGREEMENT=yes -d neo4j:4.4-enterprise
       shell: bash
-      if: ${{ runner.os != 'windows-latest' }}
+      if: ${{ runner.os != 'Windows' }}
     - name: Start Postgres Docker
       run: docker run --restart always --publish=5432:5432 --env POSTGRES_PASSWORD=postgres -d postgres:11.21-bullseye
       shell: bash
-      if: ${{ runner.os != 'windows-latest' }}
+      if: ${{ runner.os != 'Windows' }}
     #----------------------------------------------
     #      run tests and code quality checks
     #----------------------------------------------
