+++ .github/workflows/docs.yaml
@@ -5,8 +5,6 @@ on: [push, pull_request]
 jobs:
   build_docs:
     runs-on: ubuntu-latest
-    env:
-      PYTHON_VERSION: 3.10.5
     permissions: write-all
     steps:
     #----------------------------------------------
@@ -32,8 +30,6 @@ jobs:
   build_and_deploy_docs:
     if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags')
     runs-on: ubuntu-latest
-    env:
-      PYTHON_VERSION: 3.10.5
     permissions: write-all
     steps:
     #----------------------------------------------
