+++ .github/actions/test/action.yaml
@@ -22,10 +22,16 @@ runs:
     #----------------------------------------------
     #     run tests and code quality checks
     #----------------------------------------------
+    - run: |
+        source .venv/scripts/activate
+        pytest --version
+      if: runner.os == 'Windows'
+    - run: |
+        source .venv/bin/activate
+        pytest --version
+      if: runner.os != 'Windows'
     - name: Run Tests
-      run: |
-        source $VENV
-        pytest --password=your_password_here
+      run: pytest --password=your_password_here
       shell: bash
     - name: Check code quality
       uses: pre-commit/action@v3.0.0
