### 2023-10-19 e2c1c94
issue #259: add shell
