+++ .github/actions/test/action.yaml
@@ -22,20 +22,19 @@ runs:
     #----------------------------------------------
     #     run tests and code quality checks
     #----------------------------------------------
-    - run: |
+    - name: Run Tests (Windows)
+      run: |
         source .venv/scripts/activate
         pytest --version
-      shell: bash
-      if: runner.os == 'Windows'
-    - run: |
-        source .venv/bin/activate
-        pytest --version
-      shell: bash
-      if: runner.os != 'Windows'
-    - name: Run Tests
-      run: |
-        source $VENV
         pytest --password=your_password_here
       shell: bash
+      if: runner.os == 'Windows'
+    - name: Run tests (Linux and MacOS)
+      run: |
+        source .venv/bin/activate
+        pytest --version
+        pytest --password=your_password_here
+      shell: bash
+      if: runner.os != 'Windows'
     - name: Check code quality
       uses: pre-commit/action@v3.0.0
