### 2023-10-19 7086309
adjust ontology test
