### 2023-10-19 8a79ba0
replace value error with warning
