### 2023-10-19 cc3af88
check for the successful creation of biocypher class without schema
