### 2023-10-19 e5e6f3f
allow mapping to be None
