### 2023-10-19 7184e33
extend integration test to building an ontology
