+++ test/test_integration.py
@@ -1,6 +1,9 @@
 import os
 
 import pytest
+import networkx as nx
+
+from biocypher._ontology import Ontology
 
 
 @pytest.mark.parametrize("l", [4], scope="function")
@@ -39,3 +42,15 @@ def test_show_ontology_structure_kwargs(core):
 
 def test_ontology_without_schema_config(core_no_schema):
     assert core_no_schema
+
+    core_no_schema._head_ontology = {
+        "url": "http://semanticweb.cs.vu.nl/2009/11/sem/",
+        "root_node": "Core",
+        "format": "rdf",
+    }
+    core_no_schema._ontology_mapping = None
+
+    core_no_schema._get_ontology()
+
+    assert isinstance(core_no_schema._ontology, Ontology)
+    assert isinstance(core_no_schema._ontology._nx_graph, nx.DiGraph)
