### 2023-10-19 de80111
raise error when attempting to visualise without .. schema configuration
