### 2023-10-19 b042d87
Bump version: 0.5.23 → 0.5.24
