### 2023-10-19 559892d
enable defaults for log and output directory .. in config
