### 2023-10-19 ebd3333
add `as_node` function to write schema info .. directly to graph. non-functional; don't know how to make an exception for the schema_info node in the schema config (if not in there, it .. will not pass the filters)
