+++ test/test_integration.py
@@ -54,3 +54,18 @@ def test_ontology_without_schema_config(core_no_schema):
 
     assert isinstance(core_no_schema._ontology, Ontology)
     assert isinstance(core_no_schema._ontology._nx_graph, nx.DiGraph)
+
+
+@pytest.mark.parametrize("l", [4], scope="function")
+def test_write_schema_info_as_node(core, _get_nodes):
+    core.add(_get_nodes)
+
+    schema = core.write_schema_info(as_node=True)
+
+    path = os.path.join(core._output_directory, "schema_info.yaml")
+    assert os.path.exists(path)
+
+    with open(path, "r") as f:
+        schema_loaded = yaml.safe_load(f)
+
+    assert schema_loaded == schema
