### 2023-10-19 202e1d0
make the schema in schema info creation .. independent from the ontology.mapping object (leads to problems if .. used after writing the schema)
