### 2023-10-19 b7ee2bc
move output directory check and creation to the .. core class to ensure always having an output directory
