+++ biocypher/_write.py
@@ -183,7 +183,7 @@ class _BatchWriter(ABC):
             quote:
                 The quote character to use for the CSV files.
 
-            dirname:
+            output_directory:
                 Path for exporting CSV files.
 
             db_name:
@@ -1887,10 +1887,6 @@ def get_writer(
 
     dbms_config = _config(dbms)
 
-    timestamp = lambda: datetime.now().strftime("%Y%m%d%H%M%S")
-    outdir = output_directory or os.path.join("biocypher-out", timestamp())
-    outdir = os.path.abspath(outdir)
-
     writer = DBMS_TO_CLASS[dbms]
 
     if not writer:
@@ -1903,7 +1899,7 @@ def get_writer(
             delimiter=dbms_config.get("delimiter"),
             array_delimiter=dbms_config.get("array_delimiter"),
             quote=dbms_config.get("quote_character"),
-            output_directory=outdir,
+            output_directory=output_directory,
             db_name=dbms_config.get("database_name"),
             import_call_bin_prefix=dbms_config.get("import_call_bin_prefix"),
             import_call_file_prefix=dbms_config.get("import_call_file_prefix"),
