### 2023-10-19 8674521
add force method to skip ancestry when writing .. nodes
