### 2023-10-19 c45fe1e
deactivate default folders again, should be .. taken care of by moving outdir folder generation to core module
