### 2023-10-19 ad04c8d
default force value in `_write_node_data`
