### 2023-10-19 7b534d3
Bump version: 0.5.24 → 0.5.25
