+++ .github/workflows/docs.yaml
@@ -1,6 +1,9 @@
 name: Docs
 
-on: [push, pull_request]
+on:
+  push:
+    branches:
+      - main
 
 jobs:
   build_docs:
