+++ .github/workflows/publish.yaml
@@ -6,8 +6,7 @@ on:
       - pyopensci
       - main
     tags:
-      - "[0-9]+.[0-9]+.[0-9]+"
-      - "[0-9]+.[0-9]+.[0-9]+(a|b|rc|post|dev)[0-9]+"
+      - "^[0-9]+.[0-9]+.[0-9]+(-[a-z]+)?(.[0-9]+)?$"
 
 jobs:
   build_and_deploy_artifact:
