### 2023-10-20 bea694c
remove start and end of string matching
