### 2023-10-20 6ed89d6
adjust test to conftest setup, fix typo
