+++ test/test_deduplicate.py
@@ -4,7 +4,7 @@ from biocypher._create import BioCypherEdge, BioCypherNode
 from biocypher._deduplicate import Deduplicator
 
 
-@pytest.mark.parametrize("lenght", [4], scope="module")
+@pytest.mark.parametrize("length", [4], scope="module")
 def test_duplicate_nodes(_get_nodes):
     dedup = Deduplicator()
     nodes = _get_nodes
@@ -28,7 +28,7 @@ def test_duplicate_nodes(_get_nodes):
     assert "p1" in dedup.duplicate_entity_ids
 
 
-@pytest.mark.parametrize("lenght", [4], scope="module")
+@pytest.mark.parametrize("length", [4], scope="module")
 def test_get_duplicate_nodes(_get_nodes):
     dedup = Deduplicator()
     nodes = _get_nodes
@@ -56,7 +56,7 @@ def test_get_duplicate_nodes(_get_nodes):
     assert "p1" in ids
 
 
-@pytest.mark.parametrize("lenght", [4], scope="module")
+@pytest.mark.parametrize("length", [4], scope="module")
 def test_duplicate_edges(_get_edges):
     dedup = Deduplicator()
     edges = _get_edges
@@ -82,7 +82,7 @@ def test_duplicate_edges(_get_edges):
     assert ("mrel2") in dedup.duplicate_relationship_ids
 
 
-@pytest.mark.parametrize("lenght", [4], scope="module")
+@pytest.mark.parametrize("length", [4], scope="module")
 def test_get_duplicate_edges(_get_edges):
     dedup = Deduplicator()
     edges = _get_edges
