+++ tutorial/data_generator.py
@@ -72,12 +72,12 @@ class Protein(Node):
         ## random amino acid sequence
 
         # random int between 50 and 250
-        random_lenght = random.randint(50, 250)
+        random_length = random.randint(50, 250)
 
         properties["sequence"] = "".join(
             [
                 random.choice("ACDEFGHIKLMNPQRSTVWY")
-                for _ in range(random_lenght)
+                for _ in range(random_length)
             ],
         )
 
@@ -136,12 +136,12 @@ class RandomPropertyProtein(Protein):
         ## random amino acid sequence
 
         # random int between 50 and 250
-        random_lenght = random.randint(50, 250)
+        random_length = random.randint(50, 250)
 
         properties["sequence"] = "".join(
             [
                 random.choice("ACDEFGHIKLMNPQRSTVWY")
-                for _ in range(random_lenght)
+                for _ in range(random_length)
             ],
         )
 
