### 2023-10-25 94b7a9f
extend test to check import statement
