### 2023-10-25 95612aa
Bump version: 0.5.25 → 0.5.26
