### 2023-10-25 7aa842a
issue #259: make publish depend on succesful test pipeline
