+++ .github/workflows/publish.yaml
@@ -1,17 +1,18 @@
 name: Publish
 
-on:
-  push:
-    branches:
-      - pyopensci
-      - main
-    tags:
-      - '*'
+on: [push]
+  #push:
+    #branches:
+    #  - pyopensci
+    #  - main
+    #tags:
+    #  - '*'
     #   - "[0-9]+.[0-9]+.[0-9]+(-[a-z]+)?(.[0-9]+)?"
 
 jobs:
   build_and_deploy_artifact:
-    if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags')
+    #if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags')
+    needs: tests_and_code_quality
     runs-on: ubuntu-latest
     env:
       PYTHON_VERSION: 3.9
@@ -42,19 +43,19 @@ jobs:
     #----------------------------------------------
     #             upload to PyPi
     #----------------------------------------------
-    - name: Publish artifact to PyPI
-      uses: pypa/gh-action-pypi-publish@release/v1
-      with:
-        repository-url: https://test.pypi.org/legacy/
+#    - name: Publish artifact to PyPI
+#      uses: pypa/gh-action-pypi-publish@release/v1
+#      with:
+#        repository-url: https://test.pypi.org/legacy/
     #----------------------------------------------
     #           create Github release
     #----------------------------------------------
-    - name: Create Github release
-      env:
-        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-        tag: ${{ github.ref_name }}
-      run: |
-        gh release create "$tag" \
-            --repo="$GITHUB_REPOSITORY" \
-            --title="$tag" \
-            --generate-notes
+#    - name: Create Github release
+#      env:
+#        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+#        tag: ${{ github.ref_name }}
+#      run: |
+#        gh release create "$tag" \
+#            --repo="$GITHUB_REPOSITORY" \
+#            --title="$tag" \
+#            --generate-notes
