### 2023-10-25 57eb92f
issue #259: make publish depend on succesful test pipeline 2
