+++ .github/workflows/publish.yaml
@@ -1,6 +1,10 @@
 name: Publish
 
-on: [push]
+on:
+  workflow_run:
+    workflows: [Tests and code quality]
+    types:
+      - completed
   #push:
     #branches:
     #  - pyopensci
@@ -12,7 +16,6 @@ on: [push]
 jobs:
   build_and_deploy_artifact:
     #if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags')
-    needs: tests_and_code_quality
     runs-on: ubuntu-latest
     env:
       PYTHON_VERSION: 3.9
