### 2023-10-25 e6feeed
issue #259: make publish depend on succesful test pipeline 3
