### 2023-10-25 9cf7734
issue #259: make publish depend on succesful test pipeline 4
