+++ .github/workflows/publish.yaml
@@ -3,20 +3,19 @@ name: Publish
 on:
   workflow_run:
     workflows: [Tests and code quality]
-    branches: [PyOpenSci_review_part_4]
+    branches: [main]
     types:
       - completed
-  #push:
-    #branches:
-    #  - pyopensci
-    #  - main
-    #tags:
-    #  - '*'
-    #   - "[0-9]+.[0-9]+.[0-9]+(-[a-z]+)?(.[0-9]+)?"
+  push:
+    branches:
+      - main
+    tags:
+      - '*'
+    # TODO: ensure semver style - "[0-9]+.[0-9]+.[0-9]+(-[a-z]+)?(.[0-9]+)?"
 
 jobs:
   build_and_deploy_artifact:
-    #if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags')
+    if: github.event_name == 'push' && github.event.ref_type == 'tag' && github.event.workflow_run.conclusion == 'success'
     runs-on: ubuntu-latest
     env:
       PYTHON_VERSION: 3.9
@@ -47,19 +46,19 @@ jobs:
     #----------------------------------------------
     #             upload to PyPi
     #----------------------------------------------
-#    - name: Publish artifact to PyPI
-#      uses: pypa/gh-action-pypi-publish@release/v1
-#      with:
-#        repository-url: https://test.pypi.org/legacy/
+    - name: Publish artifact to PyPI
+      uses: pypa/gh-action-pypi-publish@release/v1
+      with:
+        repository-url: https://test.pypi.org/legacy/
     #----------------------------------------------
     #           create Github release
     #----------------------------------------------
-#    - name: Create Github release
-#      env:
-#        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-#        tag: ${{ github.ref_name }}
-#      run: |
-#        gh release create "$tag" \
-#            --repo="$GITHUB_REPOSITORY" \
-#            --title="$tag" \
-#            --generate-notes
+    - name: Create Github release
+      env:
+        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        tag: ${{ github.ref_name }}
+      run: |
+        gh release create "$tag" \
+            --repo="$GITHUB_REPOSITORY" \
+            --title="$tag" \
+            --generate-notes
