### 2023-10-25 e5216e0
fix github token permissions
