+++ .github/actions/test_coverage/action.yaml
@@ -3,6 +3,8 @@ description: 'Check test coverage'
 
 runs:
   using: "composite"
+  permissions:
+    contents: write
   steps:
     - name: Generate coverage report
       run: poetry run coverage run -m pytest --password=your_password_here
