+++ .github/workflows/tests_and_code_quality.yaml
@@ -1,6 +1,6 @@
 name: Tests and code quality
 
-on: [push]
+on: [push, workflow_call]
 
 jobs:
   tests_and_code_quality:
