+++ .github/workflows/publish.yaml
@@ -10,7 +10,7 @@ on:
 
 jobs:
   call-tests_and_code_quality:
-    uses: ./.github/workflows/tests_and_code_quality.yml
+    uses: "./.github/workflows/tests_and_code_quality.yaml"
 
   build_and_deploy_artifact:
     if: github.event_name == 'push' && github.event.ref_type == 'tag'
