+++ DEVELOPER.md
@@ -13,38 +13,62 @@ before starting development.
 ### Setup the environment
 
 - Install dependencies from the lock file: `poetry install`
-- Use the environment: You can either run commands directly with `poetry run <command>` or open a shell with `poetry shell` and then run commands directly.
+
+- Use the environment: You can either run commands directly with `poetry run
+<command>` or open a shell with `poetry shell` and then run commands directly.
 
 ### Updating the environment
 
 If you want to fix dependency issues, please do so in the Poetry
 framework. If Poetry does not work for you for some reason, please let us know.
 
-The Poetry dependencies are organized in groups.
-There are groups with dependencies needed for running BioCypher (`[tool.poetry.dependencies` with the group name `main`) and a group with dependencies needed for development (`[tool.poetry.group.dev.dependencies` with the group name `dev`).
+The Poetry dependencies are organized in groups. There are groups with
+dependencies needed for running BioCypher (`[tool.poetry.dependencies` with the
+group name `main`) and a group with dependencies needed for development
+(`[tool.poetry.group.dev.dependencies` with the group name `dev`).
 
 For adding new dependencies:
+
 - Add new dependencies: `poetry add <dependency> -- group <group>`
-- Update lock file (after adding new dependencies in pyproject.toml): `poetry lock`
+
+- Update lock file (after adding new dependencies in pyproject.toml): `poetry
+lock`
 
 ## Code quality and formal requirements
 
-For ensuring code quality following tools are used:
+For ensuring code quality, the following tools are used:
+
 - [isort](https://isort.readthedocs.io/en/latest/) for sorting imports
+
 - [black](https://black.readthedocs.io/en/stable/) for automated code formatting
-- [pre-commit-hooks](https://github.com/pre-commit/pre-commit-hooks) for ensuring some general rules
-- [pep585-upgrade](https://github.com/snok/pep585-upgrade) for automatically upgrading type hints to the new native types defined in PEP 585
-- [pygrep-hooks](https://github.com/pre-commit/pygrep-hooks) for ensuring some general naming rules
 
-Pre-commit hooks are used to automatically run these tools before each commit. They are defined in [.pre-commit-config.yaml](./.pre-commit-config.yaml). To install the hooks run `poetry run pre-commit install`. The hooks are then executed before each commit.
-For running the hook for all project files (not only the changed ones) run `poetry run pre-commit run --all-files`.
+- [pre-commit-hooks](https://github.com/pre-commit/pre-commit-hooks) for
+ensuring some general rules
 
-The project uses [Sphinx](https://www.sphinx-doc.org/en/master/) autodoc GitHub
-Actions workflow to generate the documentation. If you add new code, please make sure that it is documented
-accordingly and in a consistent manner with the existing code base.
-Especially the docstrings should follow the [Google style guide](https://sphinxcontrib-napoleon.readthedocs.io/en/latest/example_google.html). To check, you can run the documentation build locally by running `make html` in the `docs` directory.
+- [pep585-upgrade](https://github.com/snok/pep585-upgrade) for automatically
+upgrading type hints to the new native types defined in PEP 585
 
-When adding new code snippets to the documentation make sure, that they are automatically tested with [doctest](https://sphinx-tutorial.readthedocs.io/step-3/#testing-your-code) to ensure, that no ouutdated code snippets are part of the documentation.
+- [pygrep-hooks](https://github.com/pre-commit/pygrep-hooks) for ensuring some
+general naming rules
+
+Pre-commit hooks are used to automatically run these tools before each commit.
+They are defined in [.pre-commit-config.yaml](./.pre-commit-config.yaml). To
+install the hooks run `poetry run pre-commit install`. The hooks are then
[…]
