+++ pyproject.toml
@@ -100,7 +100,7 @@ exclude = '''
 
 [tool.isort]
 from_first = true
-line_length = 79
+line_length = 80
 multi_line_output = 3
 include_trailing_comma = true
 use_parentheses = true
@@ -120,5 +120,5 @@ per-file-ignores = [
     "tests/*:D100,D101,D102",
     "*/__init__.py:F401"
 ]
-max-line-length = 88
+max-line-length = 80
 count = true
