+++ .github/workflows/publish.yaml
@@ -13,7 +13,7 @@ jobs:
     uses: "./.github/workflows/tests_and_code_quality.yaml"
 
   build_and_deploy_artifact:
-    #if: github.event_name == 'push' && github.event.ref_type == 'tag'
+    if: github.event_name == 'push' && github.event.ref_type == 'tag'
     needs: call-tests_and_code_quality
     runs-on: ubuntu-latest
     env:
