+++ .github/workflows/docs.yaml
@@ -31,7 +31,7 @@ jobs:
       uses: ./.github/actions/build_docs
 
   build_and_deploy_docs:
-    if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags')
+    if: github.ref_type == 'tag'
     runs-on: ubuntu-latest
     permissions: write-all
     steps:
