+++ .github/workflows/publish.yaml
@@ -4,8 +4,8 @@ on:
   push:
     branches:
       - main
-    tags:
-      - '*'
+    #tags:
+    #  - '*'
     # TODO: ensure semver style - "[0-9]+.[0-9]+.[0-9]+(-[a-z]+)?(.[0-9]+)?"
 
 jobs:
@@ -13,7 +13,7 @@ jobs:
     uses: "./.github/workflows/tests_and_code_quality.yaml"
 
   build_and_deploy_artifact:
-    if: github.event_name == 'push' && github.event.ref_type == 'tag'
+    if: github.ref_type == 'tag'
     needs: call-tests_and_code_quality
     runs-on: ubuntu-latest
     env:
