+++ .github/workflows/publish.yaml
@@ -13,7 +13,7 @@ jobs:
     uses: "./.github/workflows/tests_and_code_quality.yaml"
 
   build_and_deploy_artifact:
-    if: github.ref_type == 'tag'
+    if: startsWith(github.ref, 'refs/tags/')
     needs: call-tests_and_code_quality
     runs-on: ubuntu-latest
     env:
