+++ .github/workflows/docs.yaml
@@ -4,7 +4,6 @@ on:
   push:
     branches:
       - main
-  workflow_call:
 
 jobs:
   build_docs:
