+++ .github/workflows/publish.yaml
@@ -2,24 +2,24 @@ name: Publish
 
 on:
   push:
-    branches:
-      - main
     tags:
       - "*"
     # TODO: ensure semver style - "[0-9]+.[0-9]+.[0-9]+(-[a-z]+)?(.[0-9]+)?"
 
 jobs:
   call-tests_and_code_quality:
+    if: github.ref == 'refs/heads/main'
     uses: "./.github/workflows/tests_and_code_quality.yaml"
 
   build_and_deploy_artifact:
+    if: github.ref == 'refs/heads/main'
     needs: call-tests_and_code_quality
     runs-on: ubuntu-latest
     env:
       PYTHON_VERSION: 3.9
-#    environment:
-#      name: staging # uploads to test.pypi.org
-#      url: https://test.pypi.org/project/biocypher/
+    environment:
+      name: staging # uploads to test.pypi.org
+      url: https://test.pypi.org/project/biocypher/
     permissions:
       id-token: write
       contents: write
@@ -44,22 +44,19 @@ jobs:
     #----------------------------------------------
     #             upload to PyPi
     #----------------------------------------------
-#    - name: Publish artifact to PyPI
-#      uses: pypa/gh-action-pypi-publish@release/v1
-#      with:
-#        repository-url: https://test.pypi.org/legacy/
+    - name: Publish artifact to PyPI
+      uses: pypa/gh-action-pypi-publish@release/v1
+      with:
+        repository-url: https://test.pypi.org/legacy/
     #----------------------------------------------
     #           create Github release
     #----------------------------------------------
-#    - name: Create Github release
-#      env:
-#        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-#        tag: ${{ github.ref_name }}
-#      run: |
-#        gh release create "$tag" \
-#            --repo="$GITHUB_REPOSITORY" \
-#            --title="$tag" \
-#            --generate-notes
-
-  call-build_and_deploy_docs:
-    uses: "./.github/workflows/docs.yaml"
+    - name: Create Github release
+      env:
+        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        tag: ${{ github.ref_name }}
+      run: |
+        gh release create "$tag" \
+            --repo="$GITHUB_REPOSITORY" \
+            --title="$tag" \
+            --generate-notes
