+++ .github/workflows/publish.yaml
@@ -2,17 +2,19 @@ name: Publish
 
 on:
   push:
+    branches:
+      - main
     tags:
       - "*"
     # TODO: ensure semver style - "[0-9]+.[0-9]+.[0-9]+(-[a-z]+)?(.[0-9]+)?"
 
 jobs:
   call-tests_and_code_quality:
-    if: github.ref == 'refs/heads/main'
+    if: github.event_name == 'push' && contains(github.ref, 'refs/tags/')
     uses: "./.github/workflows/tests_and_code_quality.yaml"
 
   build_and_deploy_artifact:
-    if: github.ref == 'refs/heads/main'
+    if: github.event_name == 'push' && contains(github.ref, 'refs/tags/')
     needs: call-tests_and_code_quality
     runs-on: ubuntu-latest
     env:
