### 2023-10-25 8da7bfc
issue #259: run tests before publishing
