+++ .github/actions/test_coverage/action.yaml
@@ -1,6 +1,11 @@
 name: 'Test coverage'
 description: 'Check test coverage'
 
+inputs:
+  token:
+    description: 'A Github PAT'
+    required: true
+
 runs:
   using: "composite"
   steps:
@@ -16,4 +21,4 @@ runs:
         REPO: self
         BRANCH: coverage
         FOLDER: docs/coverage
-        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        GITHUB_TOKEN: ${{ inputs.token }}
