+++ .github/workflows/publish.yaml
@@ -3,15 +3,19 @@ name: Publish
 on:
   push:
     branches:
-      - pyopensci
       - main
     tags:
-      - '*'
-    #   - "[0-9]+.[0-9]+.[0-9]+(-[a-z]+)?(.[0-9]+)?"
+      - "*"
+    # TODO: ensure semver style - "[0-9]+.[0-9]+.[0-9]+(-[a-z]+)?(.[0-9]+)?"
 
 jobs:
+  call-tests_and_code_quality:
+    if: github.event_name == 'push' && contains(github.ref, 'refs/tags/')
+    uses: "./.github/workflows/tests_and_code_quality.yaml"
+
   build_and_deploy_artifact:
-    if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags')
+    if: github.event_name == 'push' && contains(github.ref, 'refs/tags/')
+    needs: call-tests_and_code_quality
     runs-on: ubuntu-latest
     env:
       PYTHON_VERSION: 3.9
