### 2023-10-25 586c32c
issue #259: fix tests
