+++ README.md
@@ -4,7 +4,7 @@
 | --- | --- | --- | --- |
 | __License__ | [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT) | __Python__ | ![Python](https://img.shields.io/pypi/pyversions/biocypher) |
 | __Package__ | [![PyPI version](https://img.shields.io/pypi/v/biocypher)](https://img.shields.io/pypi/v/biocypher) [![Downloads](https://static.pepy.tech/badge/biocypher)](https://pepy.tech/project/biocypher) | __Build status__ | [![CI](https://github.com/biocypher/biocypher/actions/workflows/tests_and_code_quality.yaml/badge.svg)](https://github.com/biocypher/biocypher/actions/workflows/tests_and_code_quality.yaml) [![Docs build](https://github.com/biocypher/biocypher/actions/workflows/docs.yaml/badge.svg)](https://github.com/biocypher/biocypher/actions/workflows/docs.yaml) |
-| __Test coverage__ | ![Coverage](https://raw.githubusercontent.com/biocypher/biocypher/coverage/coverage.svg) | __Code style__ | [![Black](https://img.shields.io/badge/code%20style-black-000000.svg)]() |
+| __Tests__ | ![Coverage](https://raw.githubusercontent.com/biocypher/biocypher/coverage/coverage.svg) | __Code style__ | [![Black](https://img.shields.io/badge/code%20style-black-000000.svg)]() |
[…]
