### 2023-10-26 c40fd46
issue #278: fix ontology root label check
