### 2023-10-26 3b6df15
issue #226: add config param to disable writing log file to disk
