### 2023-10-30 e375f9d
issue #253: improve readability in tutorial by using lorem ipsum text
