### 2023-11-07 83fdb64
change log to disk default setting defaults are set in the config YAML
