### 2023-11-07 3c9a79b
issue #284: add unit test
