### 2023-11-07 #110 @arianesasso
Hey @slobentanzer, I just checked the link and it sounds like you addressed all the reviewers' comments! I changed the status of your submission and now we need @Midnighter and @Zethson to go over their checklists again and confirm that all the points were addressed. You are almost there 🥳 . Thank you for all the hard work!
