### 2023-11-07 d3195ea
issue #284: refactor test based on PR feedback
