+++ biocypher/_metadata.py
@@ -19,7 +19,7 @@ import importlib.metadata
 
 import toml
 
-_VERSION = "0.5.31"
+_VERSION = "0.5.32"
 
 
 def get_metadata():
