### 2023-11-08 #110 @Midnighter
Can confirm that all our comments regarding the code project have been addressed. Still think it'd be motivating for potential users unfamiliar with or new to knowledge graphs to see some usage scenarios, but I can understand that you consider it too much right now.

We recommend approval.
