### 2023-11-14 e6d6044
fix(neo4j): v4 only supports --version while v5 supports both
