### 2023-11-15 #110 @NickleDave
Hi all :wave:  I'm here to help our excellent editor @arianesasso who had a family emergency.

Thank you @Midnighter and @Zethson for confirming your comments were addressed, and letting us know you recommend approval.

@slobentanzer I am happy to report that 🎉 BioCypher has been apbutproved by pyOpenSci! Thank you for submitting! 🎉

## Author Wrap Up Tasks

- [ ] Add the badge for pyOpenSci peer-review to the README.md of <package-name-here>. The badge should be `[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.

## Editor Final Checks

---

If you have any feedback for us about the review process please feel free to share it here. We are always looking to improve our process and documentation in the
