### 2023-11-20 66efa6a
add pyopensci badge; will also generate a new .. release to enable Zenodo archiving
