### 2023-11-27 ddd5635
add list creation to API reference example
