### 2023-11-27 80376a2
remove slow test as it only works after the .. method has run. we need something that can warn while the method is .. still running, which involves threading.
