### 2023-11-28 #110 @NickleDave
Thank you @slobentanzer and congratulations again! 🎉

I have announced it on our [Discourse](https://pyopensci.discourse.group/t/welcome-biocypher-to-the-pyopensci-ecosystem/405) too.

If you are interested, we'd love to help you spread the word about BioCypher with a blog post on pyOpenSci's site.

@Midnighter and @Zethson huge thank you again for your work as reviewers. Please if you have a few minutes it would also help us if you could fill out the [post-review survey](https://forms.gle/BLGVCfUdiJHS5YJY7).

And thanks of course to @arianesasso for being the editor for this one. I'll go ahead and close this issue.
