### 2023-11-28 #110 @NickleDave
Definitely not required @slobentanzer, we just want to give you a venue to share BioCypher. Sometimes a blog post can be a way to talk about the package that's not as clear from docs and papers.
I'll invite you to our Slack team too if you'd like to join there and ask questions.
